feat(analysis): combo-verify CLI over the corpus (PR-5)
Extracts the 53-row combo corpus + board/GameAction driver toolkit out of the #[cfg(test)] module into a feature-gated analysis::corpus module (every driver parameterized on &CardDatabase), and adds the combo-verify feature, a thin combo-verify bin, and the 'cargo combo-verify' alias. The bin drives each corpus combo through the EXISTING detector (detect_loop for offline combos, the live drain reducer for the two live rows) and reports per row: confirmed / gated / deferred / failed. Zero new game logic; the engine owns all detection.

The corpus partition is locked by a shape test (driven ∪ gated ∪ deferred = 53, pairwise disjoint); classify_status and classify_live each carry a measured rubber-stamp revert-probe. Acceptance: 'cargo combo-verify' -> 12 confirmed / 4 gated / 37 deferred / 0 failed, exit 0. All 23 pre-existing corpus tests still run with identical assertions via the shared drivers.
