+++ .cargo/config.toml
@@ -12,6 +12,7 @@ coverage = "run --profile tool --bin coverage-report -- data/"
 parser-gaps = "run --profile tool --bin parser-gap-analyzer -- data/"
 rules-audit = "run --profile tool --bin rules-audit --features audit --"
 semantic-audit = "run --profile tool --features cli --bin oracle-gen -- semantic-audit"
+combo-verify = "run --profile tool --features combo-verify --bin combo-verify -- data/"
 scrape-feeds = "run --release -p feed-scraper --"
 tune-ai = "run --release --features tune --bin ai-tune --"
 ai-gate = "run --bin ai-gate --"
