+++ crates/engine/src/analysis/corpus_tests.rs
@@ -1,58 +1,38 @@
 //! Corpus harness for the infinite-combo detector (Engine A).
 //!
 //! This is the acceptance suite described in `.planning/combo-detection/`
-//! `IMPLEMENTATION.md` §8: one data row per combo, plus a soundness set. Each row
-//! names the combo's documented unbounded resource axis and expected
-//! [`WinKind`]; the detector ([`detect_loop`]) is "done" for a row when, driven
-//! over that combo's loop, it confirms the loop and names ≥1 of the expected axes
-//! with the expected `win_kind` — and emits NO certificate on a non-loop board.
+//! `IMPLEMENTATION.md` §8: one data row per combo, plus a soundness set. The corpus
+//! table and the bespoke driver toolkit live in the shared [`crate::analysis::corpus`]
+//! module (parameterized on `&CardDatabase`) so the `combo-verify` CLI and these
+//! tests drive ONE implementation; this module holds the `#[cfg(test)]` assertions
+//! and the test-only synthetic-loop helpers.
 //!
-//! # Layers, by what each asserts
+//! Three layers, by what each asserts.
 //!
-//! 1. **Driven end-to-end through the real pipeline** (`drive_*` /
-//!    `drive_combo_*` tests). Each is built from the *real* card-data export with
-//!    the cards' actual parsed abilities, its specific action cycle is driven
-//!    through `apply()` via [`LoopProbe`], and the cycle is confirmed by
-//!    [`detect_loop`] against the row's documented unbounded family + `WinKind`.
-//!    The set (see `DRIVEN_ROW_INDICES`): Heliod + Walking Ballista; Kilo,
-//!    Freed, Relic; Grim Monolith + Power Artifact; Devoted Druid + Vizier; Bloom
-//!    Tender + Freed; Priest of Titania + Umbral Mantle; Selvala + Staff of
-//!    Domination; Faeburrow + Pemmin's Aura; Marwyn + Sword of the Paruns; Sanguine
-//!    Bond + Exquisite Blood; Marauding Blight-Priest + Bloodthirsty Conqueror; Spike
-//!    Feeder + Archangel. The two drain-feedback combos are driven LIVE through the
-//!    per-beat `apply(PassPriority)` reducer (PR-3 Option C — the persisted
-//!    `loop_detect_ring` + the §3 reconcile shortcut), not the offline `detect_loop`
-//!    harness. Two synthetic loops (`drive_damage_loop_certificate`
-//!    plus the negatives `drive_board_change_is_not_a_loop` /
-//!    `drive_idle_board_is_not_a_loop`) exercise the same pipeline without the
-//!    export. These are the discriminating regression tests — reverting either
-//!    gate of `detect_loop` (board-equality or net-progress) flips an assertion,
-//!    and each `drive_combo_*` is revert-probed (omit the loop-closing action ⇒
-//!    no certificate).
+//! Driven end-to-end through the real pipeline (`drive_*` / `drive_combo_*`):
+//! each is built from the real card-data export with the cards' actual parsed
+//! abilities, its action cycle is driven through `apply()` via [`LoopProbe`], and
+//! confirmed by [`detect_loop`] against the row's documented family + `WinKind`
+//! (the driven set is `corpus::DRIVERS`). The two drain-feedback combos are driven
+//! LIVE through the per-beat `apply(PassPriority)` reducer (the persisted
+//! `loop_detect_ring` + the §3 reconcile shortcut). Two synthetic loops
+//! (`drive_damage_loop_certificate` plus the negatives) exercise the same pipeline
+//! without the export. Reverting either `detect_loop` gate flips an assertion.
 //!
-//! 2. **Corpus card-availability over ALL 53 rows**
-//!    (`corpus_cards_present_and_implementation_status_matches_gating`). Loads
-//!    every card of every combo from the real export and asserts the §3
-//!    card-support prerequisite: all present, and every non-gated combo is fully
-//!    modeled (no top-level `Unimplemented`) — i.e. genuinely *available* to
-//!    drive. Runs against the real export when present (the maintainer's local
-//!    run); skips gracefully when the gitignored export is absent (CI / fresh
-//!    checkout) so it never fails spuriously.
+//! Corpus card-availability over all 53 rows
+//! (`corpus_cards_present_and_implementation_status_matches_gating`): every card
+//! present, and every non-gated combo fully modeled (no top-level `Unimplemented`).
+//! Skips gracefully when the gitignored export is absent.
 //!
-//! 3. **Corpus table** (`CORPUS`) + shape-lock meta-test. All 53 combos as data
-//!    rows. Each undriven row needs a bespoke board install + exact `GameAction`
-//!    sequence (auras, attachments, timing windows); the ones not yet driven are
-//!    documented in the "Remaining corpus rows" block below with a PRECISE,
-//!    measured reason they cannot be confirmed on today's engine model (object
-//!    re-entry under id-comparing loop equality, extra-turn/combat re-entry,
-//!    per-color net-progress, drain cascades, card gating) — not vague TODOs. A
-//!    follow-up (or PR-5's `combo-verify` CLI) extends the projection to reach the
-//!    object-re-entry class. The detector each would exercise is already covered
-//!    by layer 1 and the `loop_check.rs` building-block tests (every `WinKind`
-//!    arm + soundness negatives).
+//! Corpus table (`corpus::CORPUS`) + shape/partition-lock meta-tests: all 53 rows
+//! partitioned into driven ∪ gated ∪ deferred. The `combo-verify` CLI
+//! ([`corpus::drive_row`]) classifies each via the same drivers;
+//! `drive_row_classifies_corpus_via_shared_pipeline` and the `classify_status`
+//! revert-probe pin that dispatch.
 
+use crate::analysis::corpus;
 use crate::analysis::resource::ResourceAxis;
-use crate::analysis::{detect_loop, LoopProbe, WinKind};
+use crate::analysis::{detect_loop, LoopCertificate, LoopProbe, WinKind};
 use crate::database::CardDatabase;
 use crate::game::scenario::{GameScenario, P0, P1};
 use crate::types::ability::{
@@ -65,495 +45,35 @@ use crate::types::mana::ManaType;
 use crate::types::phase::Phase;
 use crate::types::player::PlayerId;
 
-/// One row of the acceptance corpus: a combo, its documented unbounded resource
-/// family, the expected [`WinKind`], and (for the 4 card-gated combos) the card
-/// whose completion unblocks it.
-struct ComboRow {
-    /// Combo name (cards), for diagnostics.
-    name: &'static str,
-    /// The exact card names that make up this combo, as they appear in the
-    /// card-data export. The corpus test loads each of these from the real export
-    /// to confirm the combo is *available* (every card present + implemented).
-    cards: &'static [&'static str],
-    /// The unbounded-resource *family* the combo produces (the §12 "Category"
-    /// column). The detector must name ≥1 axis of this family.
-    family: ResourceFamily,
-    /// The expected `WinKind` once the loop is driven.
-    win_kind: WinKind,
-    /// `Some(card)` if this row is gated on an unimplemented card
-    /// (D3 / #19 / #36 / #49) — kept as a card-presence-only data row, not
-    /// driven; `None` otherwise.
-    gated_on: Option<&'static str>,
+/// The shared card database, loaded from the committed integration fixture
+/// (or the full export via `FORGE_TEST_FULL_DB=1`).
+fn card_db() -> &'static CardDatabase {
+    crate::test_support::shared_card_db()
 }
 
-/// The §12 unbounded-resource families, mapped to the concrete [`ResourceAxis`]
-/// the detector reports. Keeps the corpus table declarative.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ResourceFamily {
-    Mana,
-    Tokens,
-    Damage,
-    Drain,
-    Mill,
-    Death,
-    Landfall,
-    Draw,
-    DrawDamage,
-    Combat,
-    Turns,
-    Counters,
-    Proliferate,
-    Engine,
-}
-
-/// The full 53-row acceptance corpus: 3 driving combos + the 50 card-disjoint
-/// corpus combos from `FEASIBILITY-AND-PLAN.md` §12. The 4 `gated_on`-nonempty
-/// rows correspond to the cards with Unimplemented parts (§3).
-const CORPUS: &[ComboRow] = &[
-    // ---- 3 driving combos ----
-    ComboRow {
-        name: "Heliod, Sun-Crowned + Walking Ballista",
-        cards: &["Heliod, Sun-Crowned", "Walking Ballista"],
-        family: ResourceFamily::Damage,
-        win_kind: WinKind::LethalDamage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Kilo, Apogee Mind + Freed from the Real + Relic of Legends",
-        cards: &[
-            "Kilo, Apogee Mind",
-            "Freed from the Real",
-            "Relic of Legends",
-        ],
-        family: ResourceFamily::Proliferate,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Doc Aurlock, Grizzled Genius + Aang, Swift Savior + Appa, Steadfast Guardian",
-        cards: &[
-            "Doc Aurlock, Grizzled Genius",
-            "Aang, Swift Savior",
-            "Appa, Steadfast Guardian",
-        ],
-        family: ResourceFamily::Tokens,
-        win_kind: WinKind::Advantage,
-        gated_on: Some("Doc Aurlock, Grizzled Genius"),
-    },
-    // ---- 50 corpus combos (§12) ----
-    ComboRow {
-        name: "Basalt Monolith + Rings of Brighthearth",
-        cards: &["Basalt Monolith", "Rings of Brighthearth"],
-        family: ResourceFamily::Mana,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Grim Monolith + Power Artifact",
-        cards: &["Grim Monolith", "Power Artifact"],
-        family: ResourceFamily::Mana,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Palinchron + Deadeye Navigator",
-        cards: &["Palinchron", "Deadeye Navigator"],
-        family: ResourceFamily::Mana,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Devoted Druid + Vizier of Remedies",
-        cards: &["Devoted Druid", "Vizier of Remedies"],
-        family: ResourceFamily::Mana,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Dramatic Reversal + Isochron Scepter",
-        cards: &["Dramatic Reversal", "Isochron Scepter"],
-        family: ResourceFamily::Mana,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Pili-Pala + Grand Architect",
-        cards: &["Pili-Pala", "Grand Architect"],
-        family: ResourceFamily::Mana,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Bloom Tender + Freed from the Real",
-        cards: &["Bloom Tender", "Freed from the Real"],
-        family: ResourceFamily::Mana,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Priest of Titania + Umbral Mantle",
-        cards: &["Priest of Titania", "Umbral Mantle"],
-        family: ResourceFamily::Mana,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Dockside Extortionist + Temur Sabertooth",
-        cards: &["Dockside Extortionist", "Temur Sabertooth"],
-        family: ResourceFamily::Mana,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Selvala, Heart of the Wilds + Staff of Domination",
-        cards: &["Selvala, Heart of the Wilds", "Staff of Domination"],
-        family: ResourceFamily::Mana,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Faeburrow Elder + Pemmin's Aura",
-        cards: &["Faeburrow Elder", "Pemmin's Aura"],
-        family: ResourceFamily::Mana,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Marwyn, the Nurturer + Sword of the Paruns",
-        cards: &["Marwyn, the Nurturer", "Sword of the Paruns"],
-        family: ResourceFamily::Mana,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Heliod, Sun-Crowned + Walking Ballista [#13]",
-        cards: &["Heliod, Sun-Crowned", "Walking Ballista"],
-        family: ResourceFamily::Damage,
-        win_kind: WinKind::LethalDamage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Mikaeus, the Unhallowed + Triskelion",
-        cards: &["Mikaeus, the Unhallowed", "Triskelion"],
-        family: ResourceFamily::Damage,
-        win_kind: WinKind::LethalDamage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Sanguine Bond + Exquisite Blood",
-        cards: &["Sanguine Bond", "Exquisite Blood"],
-        family: ResourceFamily::Drain,
-        win_kind: WinKind::LethalDamage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Marauding Blight-Priest + Bloodthirsty Conqueror",
-        cards: &["Marauding Blight-Priest", "Bloodthirsty Conqueror"],
-        family: ResourceFamily::Drain,
-        win_kind: WinKind::LethalDamage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Niv-Mizzet, the Firemind + Curiosity",
-        cards: &["Niv-Mizzet, the Firemind", "Curiosity"],
-        family: ResourceFamily::DrawDamage,
-        win_kind: WinKind::LethalDamage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Blasphemous Act + Repercussion",
-        cards: &["Blasphemous Act", "Repercussion"],
-        family: ResourceFamily::Damage,
-        win_kind: WinKind::LethalDamage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Professor Onyx + Chain of Smog",
-        cards: &["Professor Onyx", "Chain of Smog"],
-        family: ResourceFamily::Drain,
-        win_kind: WinKind::LethalDamage,
-        gated_on: Some("Professor Onyx"),
-    },
-    ComboRow {
-        name: "Kiki-Jiki, Mirror Breaker + Zealous Conscripts",
-        cards: &["Kiki-Jiki, Mirror Breaker", "Zealous Conscripts"],
-        family: ResourceFamily::Tokens,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Splinter Twin + Deceiver Exarch",
-        cards: &["Splinter Twin", "Deceiver Exarch"],
-        family: ResourceFamily::Tokens,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Midnight Guard + Presence of Gond",
-        cards: &["Midnight Guard", "Presence of Gond"],
-        family: ResourceFamily::Tokens,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Scurry Oak + Ivy Lane Denizen",
-        cards: &["Scurry Oak", "Ivy Lane Denizen"],
-        family: ResourceFamily::Tokens,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Dualcaster Mage + Twinflame",
-        cards: &["Dualcaster Mage", "Twinflame"],
-        family: ResourceFamily::Tokens,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Felidar Guardian + Saheeli Rai",
-        cards: &["Felidar Guardian", "Saheeli Rai"],
-        family: ResourceFamily::Tokens,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Basking Broodscale + Rosie Cotton of South Lane",
-        cards: &["Basking Broodscale", "Rosie Cotton of South Lane"],
-        family: ResourceFamily::Tokens,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Ratadrabik of Urborg + Boromir, Warden of the Tower",
-        cards: &["Ratadrabik of Urborg", "Boromir, Warden of the Tower"],
-        family: ResourceFamily::Tokens,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Niv-Mizzet, Parun + Ophidian Eye",
-        cards: &["Niv-Mizzet, Parun", "Ophidian Eye"],
-        family: ResourceFamily::Draw,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Narset's Reversal + Twinning Staff",
-        cards: &["Narset's Reversal", "Twinning Staff"],
-        family: ResourceFamily::Draw,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Aggravated Assault + Sword of Feast and Famine",
-        cards: &["Aggravated Assault", "Sword of Feast and Famine"],
-        family: ResourceFamily::Combat,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Combat Celebrant + Helm of the Host",
-        cards: &["Combat Celebrant", "Helm of the Host"],
-        family: ResourceFamily::Combat,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Time Sieve + Thopter Assembly",
-        cards: &["Time Sieve", "Thopter Assembly"],
-        family: ResourceFamily::Turns,
-        win_kind: WinKind::ExtraTurns,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Lotus Cobra + Springheart Nantuko",
-        cards: &["Lotus Cobra", "Springheart Nantuko"],
-        family: ResourceFamily::Landfall,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Ashaya, Soul of the Wild + Quirion Ranger",
-        cards: &["Ashaya, Soul of the Wild", "Quirion Ranger"],
-        family: ResourceFamily::Landfall,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Scute Swarm + Retreat to Coralhelm",
-        cards: &["Scute Swarm", "Retreat to Coralhelm"],
-        family: ResourceFamily::Landfall,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Worldgorger Dragon + Animate Dead",
-        cards: &["Worldgorger Dragon", "Animate Dead"],
-        family: ResourceFamily::Engine,
-        win_kind: WinKind::Advantage,
-        gated_on: Some("Animate Dead"),
-    },
-    ComboRow {
-        name: "Food Chain + Eternal Scourge",
-        cards: &["Food Chain", "Eternal Scourge"],
-        family: ResourceFamily::Engine,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Tidespout Tyrant + Sol Ring",
-        cards: &["Tidespout Tyrant", "Sol Ring"],
-        family: ResourceFamily::Engine,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Aetherflux Reservoir + Bolas's Citadel + Sensei's Divining Top",
-        cards: &[
-            "Aetherflux Reservoir",
-            "Bolas's Citadel",
-            "Sensei's Divining Top",
-        ],
-        family: ResourceFamily::Damage,
-        win_kind: WinKind::LethalDamage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Abdel Adrian + Restoration Angel + Ephemerate",
-        cards: &[
-            "Abdel Adrian, Gorion's Ward",
-            "Restoration Angel",
-            "Ephemerate",
-        ],
-        family: ResourceFamily::Tokens,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Underworld Breach + Lion's Eye Diamond + Brain Freeze",
-        cards: &["Underworld Breach", "Lion's Eye Diamond", "Brain Freeze"],
-        family: ResourceFamily::Mill,
-        win_kind: WinKind::Decking,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Gravecrawler + Phyrexian Altar + Blood Artist",
-        cards: &["Gravecrawler", "Phyrexian Altar", "Blood Artist"],
-        family: ResourceFamily::Death,
-        win_kind: WinKind::LethalDamage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Karmic Guide + Reveillark + Viscera Seer",
-        cards: &["Karmic Guide", "Reveillark", "Viscera Seer"],
-        family: ResourceFamily::Death,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Chatterfang + Warren Soultrader + Academy Manufactor",
-        cards: &[
-            "Chatterfang, Squirrel General",
-            "Warren Soultrader",
-            "Academy Manufactor",
-        ],
-        family: ResourceFamily::Death,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Reassembling Skeleton + Ashnod's Altar + Nim Deathmantle",
-        cards: &["Reassembling Skeleton", "Ashnod's Altar", "Nim Deathmantle"],
-        family: ResourceFamily::Death,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Thopter Foundry + Sword of the Meek + Krark-Clan Ironworks",
-        cards: &[
-            "Thopter Foundry",
-            "Sword of the Meek",
-            "Krark-Clan Ironworks",
-        ],
-        family: ResourceFamily::Engine,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Spike Feeder + Archangel of Thune",
-        cards: &["Spike Feeder", "Archangel of Thune"],
-        family: ResourceFamily::Counters,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Earthcraft + Squirrel Nest",
-        cards: &["Earthcraft", "Squirrel Nest"],
-        family: ResourceFamily::Tokens,
-        win_kind: WinKind::Advantage,
-        gated_on: None,
-    },
-    ComboRow {
-        name: "Grindstone + Painter's Servant",
-        cards: &["Grindstone", "Painter's Servant"],
-        family: ResourceFamily::Mill,
-        win_kind: WinKind::Decking,
-        gated_on: Some("Grindstone"),
-    },
-    ComboRow {
-        name: "Helm of Obedience + Rest in Peace",
-        cards: &["Helm of Obedience", "Rest in Peace"],
-        family: ResourceFamily::Mill,
-        win_kind: WinKind::Decking,
-        gated_on: None,
-    },
-];
-
-impl ResourceFamily {
-    /// The concrete [`ResourceAxis`] (against opponent `P1` where the family is
-    /// directed at an opponent) this family is expected to name. Used by a future
-    /// per-row driver; the meta-test asserts it is total over the enum.
-    fn expected_axis(self) -> ResourceAxis {
-        match self {
-            ResourceFamily::Mana => ResourceAxis::Mana(ManaType::Colorless),
-            ResourceFamily::Tokens => ResourceAxis::TokensCreated,
-            ResourceFamily::Damage => ResourceAxis::DamageDealt(P1),
-            ResourceFamily::Drain => ResourceAxis::Life(P1),
-            ResourceFamily::Mill => ResourceAxis::LibraryDelta(P1),
-            ResourceFamily::Death => ResourceAxis::DeathTriggers,
-            ResourceFamily::Landfall => ResourceAxis::LandfallTriggers,
-            ResourceFamily::Draw => ResourceAxis::CardsDrawn,
-            ResourceFamily::DrawDamage => ResourceAxis::DamageDealt(P1),
-            ResourceFamily::Combat => ResourceAxis::CombatPhases,
-            ResourceFamily::Turns => ResourceAxis::ExtraTurns,
-            ResourceFamily::Counters => ResourceAxis::Counter(
-                crate::analysis::resource::CounterClass::Plus1Plus1,
-                crate::analysis::resource::ObjectClass::Creature,
-            ),
-            ResourceFamily::Proliferate => {
-                ResourceAxis::Trigger(crate::analysis::resource::TriggerKind::Proliferate)
-            }
-            ResourceFamily::Engine => ResourceAxis::Mana(ManaType::Colorless),
-        }
-    }
-}
+// ===========================================================================
+// Meta-tests over the shared corpus table.
+// ===========================================================================
 
 /// META-TEST: lock the corpus shape so an accidental row deletion or miscount
 /// fails loudly. 53 rows total (3 driving + 50 corpus), exactly 4 card-gated.
 #[test]
 fn corpus_table_shape_is_locked() {
     assert_eq!(
-        CORPUS.len(),
+        corpus::corpus_len(),
         53,
         "corpus must hold all 3 driving + 50 combos"
     );
-    let gated = CORPUS.iter().filter(|r| r.gated_on.is_some()).count();
+    let gated = (0..corpus::corpus_len())
+        .filter(|&i| corpus::row(i).gated_on.is_some())
+        .count();
     assert_eq!(
         gated, 4,
         "exactly 4 rows are card-gated (Doc Aurlock / Professor Onyx / Animate Dead / Grindstone)"
     );
     // Every row's expected axis must be derivable (total match over the enum).
-    for row in CORPUS {
+    for i in 0..corpus::corpus_len() {
+        let row = corpus::row(i);
         let _ = row.family.expected_axis();
         // A directed win family must classify as a loss condition, never Advantage.
         match row.win_kind {
@@ -566,14 +86,63 @@ fn corpus_table_shape_is_locked() {
         }
     }
     // 49 of 53 are testable today (gated count is the complement).
-    let testable = CORPUS.len() - gated;
+    let testable = corpus::corpus_len() - gated;
     assert_eq!(testable, 49, "49 corpus combos are testable once driven");
+}
+
+/// META-TEST: the corpus is a clean partition — every row is exactly one of
+/// {driven, gated, deferred}, pairwise disjoint, covering all 53. The driven set
+/// is `corpus::DRIVERS`, gated is `gated_on.is_some()`, deferred is
+/// `deferral.is_some()`; a driven/gated row must NOT also carry a deferral bucket.
+#[test]
+fn corpus_partition_is_locked() {
+    use std::collections::BTreeSet;
+    let n = corpus::corpus_len();
+    let driven: BTreeSet<usize> = corpus::DRIVERS.iter().map(|(i, _)| *i).collect();
+    let gated: BTreeSet<usize> = (0..n)
+        .filter(|&i| corpus::row(i).gated_on.is_some())
+        .collect();
+    let deferred: BTreeSet<usize> = (0..n)
+        .filter(|&i| corpus::row(i).deferral.is_some())
+        .collect();
+
+    assert_eq!(driven.len(), 12, "12 driven rows");
+    assert_eq!(gated.len(), 4, "4 gated rows");
+    assert_eq!(deferred.len(), 37, "37 deferred rows");
+
+    assert!(driven.is_disjoint(&gated), "driven ∩ gated must be empty");
+    assert!(
+        driven.is_disjoint(&deferred),
+        "driven ∩ deferred must be empty"
+    );
+    assert!(
+        gated.is_disjoint(&deferred),
+        "gated ∩ deferred must be empty"
+    );
+
+    let mut union = BTreeSet::new();
+    union.extend(driven.iter().copied());
+    union.extend(gated.iter().copied());
+    union.extend(deferred.iter().copied());
+    assert_eq!(
+        union.len(),
+        n,
+        "driven ∪ gated ∪ deferred must cover every one of the {n} rows"
+    );
+    assert_eq!(n, 53);
+
+    // Exclusivity: a driven or gated row must not also declare a deferral bucket.
+    for &i in driven.iter().chain(gated.iter()) {
+        assert!(
+            corpus::row(i).deferral.is_none(),
+            "row {i} is driven/gated and must not also declare a DeferralBucket"
+        );
+    }
 }
 
 /// True if any top-level ability/trigger/static/replacement of `face` parsed to
 /// `Effect::Unimplemented` — i.e. the card is not yet fully modeled.
 fn face_has_unimplemented(face: &crate::types::card::CardFace) -> bool {
-    use crate::types::ability::Effect;
     let ability_unimpl = |def: &AbilityDefinition| {
         let mut stack = vec![&*def.effect];
         while let Some(e) = stack.pop() {
@@ -593,14 +162,8 @@ fn face_has_unimplemented(face: &crate::types::card::CardFace) -> bool {
 /// ACCEPTANCE OVER THE WHOLE CORPUS (all 53 rows): every card of every combo is
 /// present in the real card-data export, and its implementation status matches
 /// the row's `gated_on` — a non-gated combo has zero `Effect::Unimplemented`
-/// across all its cards (so it is genuinely *available* to drive), while a gated
-/// combo legitimately contains an unmodeled card. This is the §3 card-support
-/// prerequisite encoded as a single data-driven test over the entire corpus.
-///
-/// Runs against the real export when present (the case the maintainer drives
-/// locally); skips gracefully when the gitignored export is absent (CI / fresh
-/// checkout) so it never fails spuriously. When it runs, it exercises ALL 49
-/// non-gated combos plus confirms the 4 gated ones are correctly classified.
+/// across all its cards, while a gated combo legitimately contains an unmodeled
+/// card. Runs against the real export when present; skips gracefully when absent.
 #[test]
 fn corpus_cards_present_and_implementation_status_matches_gating() {
     let db = card_db();
@@ -610,17 +173,17 @@ fn corpus_cards_present_and_implementation_status_matches_gating() {
     // that would silently make a "testable" row undriveable).
     let mut unexpected_unimpl: Vec<String> = Vec::new();
 
-    for row in CORPUS {
+    for i in 0..corpus::corpus_len() {
+        let row = corpus::row(i);
         for &card in row.cards {
             match db.get_face_by_name(card) {
                 None => missing.push(format!("{} (in {})", card, row.name)),
                 Some(face) => {
                     // Only the non-gated rows must be fully modeled; the 4 gated
-                    // rows legitimately contain an unmodeled card (§3). A nested
+                    // rows legitimately contain an unmodeled card. A nested
                     // Unimplemented in a cost/replacement may not be surfaced by
                     // `face_has_unimplemented` (it walks top-level ability/trigger
-                    // effects), so this is a conservative *non-regression* check:
-                    // a top-level Unimplemented on a "testable" combo is a defect.
+                    // effects), so this is a conservative *non-regression* check.
                     if row.gated_on.is_none() && face_has_unimplemented(face) {
                         unexpected_unimpl.push(format!("{} (in {})", card, row.name));
                     }
@@ -639,11 +202,354 @@ fn corpus_cards_present_and_implementation_status_matches_gating() {
     );
 }
 
+/// Meta: the driven set is a subset of the non-gated corpus and every driven row
+/// is currently driven by a real driver (kept honest as drivers are added/removed
+/// in `corpus::DRIVERS`).
+#[test]
+fn confirmed_drivers_match_expected() {
+    for &(idx, _) in corpus::DRIVERS {
+        assert!(idx < corpus::corpus_len(), "driven index in range");
+        assert!(
+            corpus::row(idx).gated_on.is_none(),
+            "{}: a driven combo must not be card-gated",
+            corpus::row(idx).name
+        );
+    }
+}
+
+// ===========================================================================
+// CLI dispatch tests: the shared `corpus::drive_row` classifier.
+// ===========================================================================
+
+/// The shared CLI dispatch ([`corpus::drive_row`]) classifies each status class
+/// through the REAL pipeline. NOTE: these are ARRAY indices; the gated rows are
+/// 2 / 21 / 38 / 51 (Doc Aurlock / Professor Onyx / Animate Dead / Grindstone) —
+/// NOT the ComboRow doc-comment's FEASIBILITY numbering (array idx 49 is a
+/// CONFIRMED driver, Spike Feeder + Archangel, not a gated row).
+#[test]
+fn drive_row_classifies_corpus_via_shared_pipeline() {
+    let db = card_db();
+
+    // Confirmed (offline): Devoted Druid + Vizier names a mana axis, Advantage.
+    match corpus::drive_row(db, 6).status {
+        corpus::RowStatus::Confirmed {
+            unbounded,
+            win_kind,
+        } => {
+            assert_eq!(win_kind, WinKind::Advantage);
+            assert!(
+                unbounded.iter().any(|a| matches!(a, ResourceAxis::Mana(_))),
+                "Devoted/Vizier must name a mana axis (got {unbounded:?})"
+            );
+        }
+        other => panic!("idx 6 must be Confirmed, got {other:?}"),
+    }
+
+    // Confirmed (live drain): both drain cascades win LethalDamage via the live path.
+    for idx in [17usize, 18] {
+        match corpus::drive_row(db, idx).status {
+            corpus::RowStatus::Confirmed { win_kind, .. } => {
+                assert_eq!(win_kind, WinKind::LethalDamage)
+            }
+            other => panic!("idx {idx} (live drain) must be Confirmed LethalDamage, got {other:?}"),
+        }
+    }
+
+    // Gated: never Failed — gated is expected. (ARRAY indices, see the note above.)
+    for (idx, card) in [
+        (2usize, "Doc Aurlock, Grizzled Genius"),
+        (21, "Professor Onyx"),
+        (38, "Animate Dead"),
+        (51, "Grindstone"),
+    ] {
+        match corpus::drive_row(db, idx).status {
+            corpus::RowStatus::Gated { card: c } => assert_eq!(c, card),
+            other => panic!("idx {idx} must be Gated (never Failed), got {other:?}"),
+        }
+    }
+
+    // Deferred: a non-driven testable row reports its structural bucket, never Failed.
+    match corpus::drive_row(db, 22).status {
+        corpus::RowStatus::Deferred { bucket } => {
+            assert_eq!(bucket, corpus::DeferralBucket::ObjectReentry)
+        }
+        other => panic!("idx 22 (Kiki-Jiki) must be Deferred(ObjectReentry), got {other:?}"),
+    }
+}
+
+/// REVERT-PROBE (the discriminating core): `classify_status` compares the driven
+/// certificate against the row spec; it must NOT rubber-stamp any `Some(cert)`.
+/// Each `Failed` assertion below flips to `Confirmed` under a specific reverted
+/// predicate (named inline), and the positive control proves the `Failed` cases
+/// are discriminating rather than vacuously always-Failed.
+#[test]
+fn classify_status_compares_against_spec_not_rubber_stamp() {
+    let row = corpus::row(6); // Devoted Druid + Vizier: family = Mana, win_kind = Advantage.
+
+    // Wrong win_kind ⇒ Failed. Revert "classify any Some(cert) as Confirmed" flips this.
+    let wrong_win = LoopCertificate {
+        unbounded: vec![ResourceAxis::DamageDealt(P1)],
+        win_kind: WinKind::LethalDamage,
+        mandatory: false,
+    };
+    assert!(
+        matches!(
+            corpus::classify_status(row, Some(wrong_win)),
+            corpus::RowStatus::Failed { .. }
+        ),
+        "a wrong-win_kind certificate must be Failed"
+    );
+
+    // Right win_kind but wrong family ⇒ Failed. Revert "drop the family check" flips this.
+    let wrong_family = LoopCertificate {
+        unbounded: vec![ResourceAxis::TokensCreated],
+        win_kind: WinKind::Advantage,
+        mandatory: false,
+    };
+    assert!(
+        matches!(
+            corpus::classify_status(row, Some(wrong_family)),
+            corpus::RowStatus::Failed { .. }
+        ),
+        "a right-win_kind / wrong-family certificate must be Failed"
+    );
+
+    // No certificate ⇒ Failed. Revert "treat None as Confirmed" flips this.
+    assert!(
+        matches!(
+            corpus::classify_status(row, None),
+            corpus::RowStatus::Failed { .. }
+        ),
+        "no certificate must be Failed"
+    );
+
+    // POSITIVE CONTROL: a correct (mana + Advantage) certificate Confirms — proving
+    // the three Failed assertions above are discriminating, not always-Failed.
+    let right = LoopCertificate {
+        unbounded: vec![ResourceAxis::Mana(ManaType::Green)],
+        win_kind: WinKind::Advantage,
+        mandatory: false,
+    };
+    assert!(
+        matches!(
+            corpus::classify_status(row, Some(right)),
+            corpus::RowStatus::Confirmed { .. }
+        ),
+        "a correct certificate must Confirm"
+    );
+}
+
+/// REVERT-PROBE (live path), symmetric with the offline probe above:
+/// `classify_live` must compare the live `GameOver` winner against the loop's
+/// controller (P0), NOT rubber-stamp any `Some(_)`. Reverting `classify_live` to
+/// `Some(_) => Confirmed` flips the wrong-winner assertion below; the `None`
+/// assertion flips under a `None => Confirmed` revert; the positive control proves
+/// the `Failed` cases are discriminating rather than vacuously always-Failed.
+#[test]
+fn classify_live_compares_winner_not_rubber_stamp() {
+    let row = corpus::row(18); // Marauding Blight-Priest + Conqueror: Drain / LethalDamage.
+
+    // Wrong winner (P1, the victim) ⇒ Failed. Revert "Some(_) => Confirmed" flips this.
+    assert!(
+        matches!(
+            corpus::classify_live(row, Some((6, P1))),
+            corpus::RowStatus::Failed { .. }
+        ),
+        "a live drain won by the wrong player (P1, the victim) must be Failed"
+    );
+
+    // No GameOver ⇒ Failed. Revert "treat None as Confirmed" flips this.
+    assert!(
+        matches!(
+            corpus::classify_live(row, None),
+            corpus::RowStatus::Failed { .. }
+        ),
+        "no GameOver must be Failed"
+    );
+
+    // POSITIVE CONTROL: the controller P0 winning a LethalDamage drain Confirms —
+    // proving the Failed assertions above are discriminating, not always-Failed.
+    match corpus::classify_live(row, Some((6, P0))) {
+        corpus::RowStatus::Confirmed { win_kind, .. } => {
+            assert_eq!(win_kind, WinKind::LethalDamage)
+        }
+        other => panic!("controller (P0) winning must Confirm LethalDamage, got {other:?}"),
+    }
+}
+
+// ===========================================================================
+// Per-combo driven acceptance tests (thin wrappers over the shared drivers).
+// Each asserts a confirmed `LoopCertificate` of the documented family + win_kind
+// via `assert_combo`, and skips (returns early) if the card-data export is absent.
+// ===========================================================================
+
+/// Assert a combo's driven certificate names the row's expected resource family
+/// and classifies the expected `win_kind`.
+fn assert_combo(idx: usize, cert: &LoopCertificate) {
+    let row = corpus::row(idx);
+    assert!(
+        cert.unbounded
+            .iter()
+            .any(|a| corpus::family_matches_axis(row.family, a)),
+        "{}: certificate {:?} must name a {:?}-family axis",
+        row.name,
+        cert.unbounded,
+        row.family,
+    );
+    assert_eq!(cert.win_kind, row.win_kind, "{}: win_kind", row.name);
+}
+
+/// HELIOD, SUN-CROWNED + WALKING BALLISTA — the canonical driving combo.
+/// DISCRIMINATION: the `expect` flips if either `detect_loop` gate is reverted.
+#[test]
+fn drive_heliod_ballista_certificate() {
+    let cert = corpus::drive_offline_heliod_ballista(card_db()).expect(
+        "Heliod + Ballista must be confirmed: board identical modulo life/damage, +1 damage/cycle",
+    );
+    assert_eq!(cert.win_kind, WinKind::LethalDamage);
+    assert!(
+        cert.covers(&[ResourceAxis::DamageDealt(P1)]),
+        "certificate must name unbounded damage to the opponent (got {:?})",
+        cert.unbounded
+    );
+}
+
+/// #4 DEVOTED DRUID + VIZIER OF REMEDIES — infinite green mana.
+#[test]
+fn drive_combo_04_devoted_vizier() {
+    let cert = corpus::drive_offline_devoted_vizier(card_db())
+        .expect("Devoted Druid + Vizier must confirm infinite green mana");
+    assert_combo(6, &cert);
+}
+
+/// #2 GRIM MONOLITH + POWER ARTIFACT — infinite colorless mana.
+#[test]
+fn drive_combo_02_grim_power() {
+    let cert = corpus::drive_offline_grim_power(card_db())
+        .expect("Grim Monolith + Power Artifact must confirm infinite colorless mana");
+    assert_combo(4, &cert);
+}
+
+/// #47 SPIKE FEEDER + ARCHANGEL OF THUNE — infinite +1/+1 counters + life.
+#[test]
+fn drive_combo_47_spike_archangel() {
+    let cert = corpus::drive_offline_spike_archangel(card_db())
+        .expect("Spike Feeder + Archangel must confirm infinite counters + life");
+    assert_combo(49, &cert);
+}
+
+/// #7 BLOOM TENDER + FREED FROM THE REAL — infinite mana.
+#[test]
+fn drive_combo_07_bloom_freed() {
+    let cert = corpus::drive_offline_bloom_freed(card_db())
+        .expect("Bloom Tender + Freed must confirm infinite mana");
+    assert_combo(9, &cert);
+}
+
+/// #11 FAEBURROW ELDER + PEMMIN'S AURA — infinite mana.
+#[test]
+fn drive_combo_11_faeburrow_pemmin() {
+    let cert = corpus::drive_offline_faeburrow_pemmin(card_db())
+        .expect("Faeburrow Elder + Pemmin's Aura must confirm infinite mana");
+    assert_combo(13, &cert);
+}
+
+/// #11 SELVALA, HEART OF THE WILDS + STAFF OF DOMINATION — infinite mana.
+#[test]
+fn drive_combo_11_selvala_staff() {
+    let cert = corpus::drive_offline_selvala_staff(card_db())
+        .expect("Selvala + Staff of Domination must confirm infinite mana");
+    assert_combo(12, &cert);
+}
+
+/// D2 KILO, APOGEE MIND + FREED + RELIC OF LEGENDS — infinite proliferate triggers.
+#[test]
+fn drive_combo_d2_kilo_freed_relic() {
+    let cert = corpus::drive_offline_kilo_freed_relic(card_db())
+        .expect("Kilo + Freed + Relic must confirm infinite proliferate triggers");
+    assert_combo(1, &cert);
+}
+
+/// #10 PRIEST OF TITANIA + UMBRAL MANTLE — infinite green mana.
+#[test]
+fn drive_combo_10_priest_umbral() {
+    let cert = corpus::drive_offline_priest_umbral(card_db())
+        .expect("Priest of Titania + Umbral Mantle must confirm infinite green mana");
+    assert_combo(10, &cert);
+}
+
+/// REVERT-PROBE for [`drive_combo_10_priest_umbral`]: omit the Umbral untap step ⇒
+/// Priest stays tapped after the first cycle, so the second cycle's tap-for-mana
+/// fails and the board is NOT identical ⇒ `run_combo` finds no certificate.
+#[test]
+fn drive_combo_10_priest_umbral_requires_untap() {
+    let mut board = corpus::build_board_green(card_db(), corpus::row(10).cards)
+        .expect("Priest of Titania board must build from the fixture");
+    let priest = board.ids[0];
+    let umbral = board.ids[1];
+    corpus::seed_subtype_creatures(board.runner.state_mut(), "Elf", 4);
+    corpus::attach_aura(board.runner.state_mut(), umbral, priest);
+    let cert = corpus::run_combo(board, |probe| {
+        if let Some(tap_idx) =
+            corpus::ability_index_where(probe.runner().state(), priest, corpus::is_mana_effect)
+        {
+            corpus::activate_and_resolve(probe, priest, tap_idx, None);
+        }
+        // No untap: Priest stays tapped, so this is not a repeatable loop.
+    });
+    assert!(
+        cert.is_none(),
+        "without the Umbral untap, Priest stays tapped — no loop"
+    );
+}
+
+/// #14 MARWYN, THE NURTURER + SWORD OF THE PARUNS — infinite green mana.
+#[test]
+fn drive_combo_14_marwyn_sword() {
+    let cert = corpus::drive_offline_marwyn_sword(card_db())
+        .expect("Marwyn + Sword of the Paruns must confirm infinite green mana");
+    assert_combo(14, &cert);
+}
+
+/// REVERT-PROBE for [`drive_combo_14_marwyn_sword`]: omit the Sword untap step ⇒
+/// Marwyn stays tapped, the next cycle's tap-for-mana fails, and the board is not
+/// identical ⇒ no certificate. Proves the modal untap is load-bearing.
+#[test]
+fn drive_combo_14_marwyn_sword_requires_untap() {
+    let mut board = corpus::build_board_green(card_db(), corpus::row(14).cards)
+        .expect("Marwyn board must build from the fixture");
+    let marwyn = board.ids[0];
+    let sword = board.ids[1];
+    {
+        let state = board.runner.state_mut();
+        if let Some(o) = state.objects.get_mut(&marwyn) {
+            o.counters
+                .insert(crate::types::counter::CounterType::Plus1Plus1, 6);
+        }
+        corpus::settle_layers(state);
+    }
+    corpus::attach_aura(board.runner.state_mut(), sword, marwyn);
+    let cert = corpus::run_combo(board, |probe| {
+        if let Some(tap_idx) =
+            corpus::ability_index_where(probe.runner().state(), marwyn, corpus::is_mana_effect)
+        {
+            corpus::activate_and_resolve(probe, marwyn, tap_idx, None);
+        }
+        // No untap: Marwyn stays tapped, so this is not a repeatable loop.
+    });
+    assert!(
+        cert.is_none(),
+        "without the Sword untap, Marwyn stays tapped — no loop"
+    );
+}
+
+// ===========================================================================
+// Synthetic loop pipeline tests (no export): board-neutral pinger + negatives.
+// ===========================================================================
+
 /// Build a battlefield creature with a board-neutral repeatable activated
 /// ability (no cost) that deals `amount` damage to any target. Each activation
-/// returns the board to an identical configuration (nothing consumed) while
-/// pumping the damage axis — a faithful net-progress loop the real `apply()`
-/// pipeline drives end-to-end. Returns the creature's `ObjectId`.
+/// returns the board to an identical configuration while pumping the damage axis.
 fn pinger_scenario(amount: i32) -> (GameScenario, ObjectId) {
     let mut scenario = GameScenario::new();
     scenario.at_phase(Phase::PreCombatMain);
@@ -697,1072 +603,12 @@ fn drive_one_ping(probe: &mut LoopProbe, pinger: ObjectId) {
     }
 }
 
-// ---------------------------------------------------------------------------
-// Real-card infrastructure: build combo boards from the actual parsed card-data
-// export, so a driven loop exercises the cards' real abilities (not synthetic
-// stand-ins). The export (`client/public/card-data.json`) is gitignored and may
-// be absent in a fresh checkout / CI; helpers that need it return `None` so the
-// caller can skip gracefully rather than fail spuriously.
-// ---------------------------------------------------------------------------
-
-/// The shared card database, loaded from the committed integration fixture
-/// (or the full export via `FORGE_TEST_FULL_DB=1`).
-fn card_db() -> &'static CardDatabase {
-    crate::test_support::shared_card_db()
-}
-
-/// Instantiate a real card by name directly onto `player`'s battlefield, with its
-/// abilities/triggers/statics parsed from the export. Already-resolved (not
-/// summoning-sick), so its activated abilities are usable the same turn. Returns
-/// the new object's id, or `None` if the card is absent from the export.
-fn install_on_battlefield(
-    state: &mut GameState,
-    db: &CardDatabase,
-    name: &str,
-    player: crate::types::player::PlayerId,
-) -> Option<ObjectId> {
-    use crate::game::printed_cards::apply_card_face_to_object;
-    use crate::types::identifiers::CardId;
-    use crate::types::zones::Zone;
-
-    let face = db.get_face_by_name(name)?;
-    let card_id = CardId(state.next_object_id);
-    let id = crate::game::zones::create_object(
-        state,
-        card_id,
-        player,
-        name.to_string(),
-        Zone::Battlefield,
-    );
-    let ts = state.next_timestamp();
-    {
-        let obj = state.objects.get_mut(&id)?;
-        apply_card_face_to_object(obj, face);
-        // CR 302.6: a pre-existing battlefield permanent is not summoning-sick.
-        obj.summoning_sick = false;
-        obj.entered_battlefield_turn = Some(state.turn_number.saturating_sub(1));
-        obj.timestamp = ts;
-    }
-    // CR 603.6: index the installed object's triggers so they fire during play.
-    crate::game::trigger_index::reindex_object_triggers(state, id);
-    Some(id)
-}
-
-// ---------------------------------------------------------------------------
-// Shared bespoke-driver toolkit. Each corpus combo is an intricate, *specific*
-// multi-action cycle, so it gets a small hand-written driver built from these
-// primitives (a generic explorer cannot sequence them). The toolkit installs the
-// real cards, floats mana, and gives uniform "activate ability / resolve /
-// answer prompt" steps so each per-combo driver stays short and readable.
-// ---------------------------------------------------------------------------
-
-/// Outcome of installing a combo: the runner plus the installed permanents in the
-/// order their card names were given.
-struct ComboBoard {
-    runner: crate::game::scenario::GameRunner,
-    ids: Vec<ObjectId>,
-}
-
-/// Build a board with the named permanents installed on P0's battlefield, a large
-/// finite mana pool floated (so mana-cost abilities can pay, while a mana-GAIN
-/// axis is still measurable — not `debug_infinite_mana`), and layers settled.
-/// `None` if the export is absent or any name is missing. Auras are installed but
-/// NOT auto-attached (each driver attaches them to the correct host).
-fn build_board(cards: &[&str]) -> ComboBoard {
-    let db = card_db();
-    let mut scenario = GameScenario::new();
-    scenario.at_phase(Phase::PreCombatMain);
-    scenario.with_life(P0, 40);
-    scenario.with_life(P1, 40);
-    let mut runner = scenario.build();
-    {
-        let state = runner.state_mut();
-        state.active_player = P0;
-        state.priority_player = P0;
-    }
-    let mut ids = Vec::new();
-    {
-        let state = runner.state_mut();
-        for &name in cards {
-            ids.push(
-                install_on_battlefield(state, db, name, P0)
-                    .expect("corpus card must be present in the committed fixture"),
-            );
-        }
-        float_mana(state, 500);
-        settle_layers(state);
-    }
-    ComboBoard { runner, ids }
-}
-
-/// Like [`build_board`], but floats GREEN-ONLY into P0's pool (no WUBRG+C pool).
-/// For a green producer whose untap/activation costs are generic ({3} etc.), the
-/// generic must be paid from the producer's own color so no per-color axis goes
-/// net-negative: a floated colorless pool (CR 106.1/106.4) would be drained by the
-/// generic costs and never replenished (these producers can't make colorless),
-/// reading as a spurious per-color deficit the detector's `is_progress` rightly
-/// rejects. Same trick `build_board_with_vanilla` uses for Selvala. Returns the
-/// post-`build()` board with the named permanents in card order.
-fn build_board_green(cards: &[&str]) -> ComboBoard {
-    let mut board = build_board(cards);
-    {
-        let state = board.runner.state_mut();
-        // CR 106.4: replace the WUBRG+C pool floated by `build_board` with a
-        // green-only pool so the producer's generic costs draw from green.
-        state.players[0].mana_pool.clear();
-        float_single_color(state, ManaType::Green, 500);
-        settle_layers(state);
-    }
-    board
-}
-
-/// Like [`build_board`], but first places a vanilla `power`/`toughness` creature
-/// on P0's battlefield (so a power-scaling mana producer like Selvala reads a high
-/// X). The vanilla creature is installed BEFORE the named combo cards, so the
-/// combo-card ids are still `ids[0..]` in card order. Returns the board plus the
-/// vanilla creature's id appended LAST in `ids`.
-fn build_board_with_vanilla(cards: &[&str], power: i32, toughness: i32) -> ComboBoard {
-    let db = card_db();
-    let mut scenario = GameScenario::new();
-    scenario.at_phase(Phase::PreCombatMain);
-    scenario.with_life(P0, 40);
-    scenario.with_life(P1, 40);
-    // CR 208.2: a high-power vanilla so a "greatest power among creatures you
-    // control" producer reads a large X (the combo's documented prerequisite).
-    let vanilla = scenario.add_vanilla(P0, power, toughness);
-    let mut runner = scenario.build();
-    {
-        let state = runner.state_mut();
-        state.active_player = P0;
-        state.priority_player = P0;
-    }
-    let mut ids = Vec::new();
-    {
-        let state = runner.state_mut();
-        for &name in cards {
-            ids.push(
-                install_on_battlefield(state, db, name, P0)
-                    .expect("corpus card must be present in the committed fixture"),
-            );
-        }
-        ids.push(vanilla);
-        // Float GREEN only (not a full WUBRG+C pool): Selvala produces green and
-        // its untap-chain costs are generic, so paying generic from green keeps
-        // every per-color axis ≥ 0. A floated colorless pool would be consumed by
-        // the generic costs and never replenished (Selvala can't make colorless),
-        // a spurious per-color deficit that the detector rightly rejects.
-        float_single_color(state, ManaType::Green, 500);
-        settle_layers(state);
-    }
-    ComboBoard { runner, ids }
-}
-
-/// Float `n` mana of a single `color` into P0's pool from a sentinel source.
-/// Used by combos whose producer makes a specific color and whose costs are
-/// generic, so paying generic from the same color avoids a spurious per-color
-/// deficit (the detector's `is_progress` rejects any net-negative color).
-fn float_single_color(state: &mut GameState, color: ManaType, n: usize) {
-    for _ in 0..n {
-        state.players[0]
-            .mana_pool
-            .add(crate::types::mana::ManaUnit::new(
-                color,
-                ObjectId(0),
-                false,
-                Vec::new(),
-            ));
-    }
-}
-
-/// Float `n` of each WUBRG+C mana into P0's pool from a sentinel source.
-fn float_mana(state: &mut GameState, n: usize) {
-    for color in [
-        ManaType::White,
-        ManaType::Blue,
-        ManaType::Black,
-        ManaType::Red,
-        ManaType::Green,
-        ManaType::Colorless,
-    ] {
-        for _ in 0..n {
-            state.players[0]
-                .mana_pool
-                .add(crate::types::mana::ManaUnit::new(
-                    color,
-                    ObjectId(0),
-                    false,
-                    Vec::new(),
-                ));
-        }
-    }
-}
-
-/// CR 613: mark layers dirty and recompute so granted keywords / aura effects /
-/// counter-derived P/T apply before the loop is driven.
-fn settle_layers(state: &mut GameState) {
-    state.layers_dirty.mark_full();
-    crate::game::layers::evaluate_layers(state);
-}
-
-/// Attach `aura` to `host` (CR 303.4): set both sides of the relationship and
-/// re-settle layers so the aura's static/granted effects apply.
-fn attach_aura(state: &mut GameState, aura: ObjectId, host: ObjectId) {
-    if let Some(o) = state.objects.get_mut(&aura) {
-        o.attached_to = Some(crate::game::game_object::AttachTarget::Object(host));
-    }
-    if let Some(h) = state.objects.get_mut(&host) {
-        if !h.attachments.contains(&aura) {
-            h.attachments.push(aura);
-        }
-    }
-    settle_layers(state);
-}
-
-/// Index of `source`'s first ability whose effect matches `pred`. Reads the live
-/// (post-layer) ability list so a granted ability is found too.
-fn ability_index_where(
-    state: &GameState,
-    source: ObjectId,
-    pred: impl Fn(&crate::types::ability::Effect) -> bool,
-) -> Option<usize> {
-    state
-        .objects
-        .get(&source)?
-        .abilities
-        .iter()
-        .position(|a| pred(&a.effect))
-}
-
-/// Activate `source`'s ability `index`, then resolve the whole stack to a clean
-/// priority window, answering any prompt by choosing `prefer_target` (or the
-/// first legal target). Returns `false` if the activation is rejected.
-fn activate_and_resolve(
-    probe: &mut LoopProbe,
-    source: ObjectId,
-    index: usize,
-    prefer_target: Option<TargetRef>,
-) -> bool {
-    if probe
-        .act(GameAction::ActivateAbility {
-            source_id: source,
-            ability_index: index,
-        })
-        .is_err()
-    {
-        return false;
-    }
-    resolve_to_priority(probe, prefer_target);
-    true
-}
-
-/// Drive the stack to a clean priority window, auto-answering target / X prompts.
-/// Prefers `prefer_target` when it is a currently-legal target; otherwise the
-/// first legal target. Bounded so a stuck state can't hang the test.
-fn resolve_to_priority(probe: &mut LoopProbe, prefer_target: Option<TargetRef>) {
-    for _ in 0..32 {
-        match &probe.runner().state().waiting_for {
-            WaitingFor::Priority { .. } if probe.runner().state().stack.is_empty() => break,
-            WaitingFor::Priority { .. } => {
-                if probe.act(GameAction::PassPriority).is_err() {
-                    break;
-                }
-            }
-            WaitingFor::TargetSelection { selection, .. }
-            | WaitingFor::TriggerTargetSelection { selection, .. } => {
-                let legal = &selection.current_legal_targets;
-                let pick = prefer_target
-                    .clone()
-                    .filter(|t| legal.contains(t))
-                    .or_else(|| legal.first().cloned());
-                let action = match pick {
-                    Some(t) => GameAction::ChooseTarget { target: Some(t) },
-                    None => GameAction::ChooseTarget { target: None },
-                };
-                if probe.act(action).is_err() {
-                    break;
-                }
-            }
-            WaitingFor::ChooseXValue { .. } => {
-                if probe.act(GameAction::ChooseX { value: 1 }).is_err() {
-                    break;
-                }
-            }
-            WaitingFor::ChooseManaColor { choice, .. } => {
-                use crate::types::game_state::{ManaChoice, ManaChoicePrompt};
-                // Answer must MATCH the prompt shape, or the engine rejects it and
-                // the state stays stuck (the bug a single-color answer hit for an
-                // X-mana `AnyCombination` producer like Selvala). Pick Blue for
-                // each unit (Blue pays {U} untap costs; with a large floated pool
-                // the exact color is otherwise immaterial).
-                let answer = match choice {
-                    ManaChoicePrompt::SingleColor { .. } => GameAction::ChooseManaColor {
-                        choice: ManaChoice::SingleColor(ManaType::Blue),
-                        count: 1,
-                    },
-                    // An X-mana "any combination" producer wants one color per
-                    // produced unit — answer with `count` Greens. Green is chosen
-                    // (not Blue) so a green-cost producer like Selvala re-pays its
-                    // own {G} from its production and the generic untap-chain costs
-                    // draw from the same green surplus, keeping every per-color
-                    // axis ≥ 0 (the detector rejects any color that goes
-                    // net-negative).
-                    ManaChoicePrompt::AnyCombination { count, .. } => GameAction::ChooseManaColor {
-                        choice: ManaChoice::Combination(vec![ManaType::Green; *count]),
-                        count: 1,
-                    },
-                    // Filter-land "pick one complete combination": take the first.
-                    ManaChoicePrompt::Combination { options } => {
-                        let combo = options.first().cloned().unwrap_or_default();
-                        GameAction::ChooseManaColor {
-                            choice: ManaChoice::Combination(combo),
-                            count: 1,
-                        }
-                    }
-                };
-                if probe.act(answer).is_err() {
-                    break;
-                }
-            }
-            WaitingFor::PayCost { choices, count, .. } => {
-                // Choose `count` objects to pay the cost (tap-creatures /
-                // sacrifice / exile). Prefer `prefer_target`'s object if it is a
-                // legal choice (so a "tap a creature" cost taps the intended one).
-                let want = match &prefer_target {
-                    Some(TargetRef::Object(o)) if choices.contains(o) => Some(*o),
-                    _ => None,
-                };
-                let mut chosen: Vec<ObjectId> = Vec::new();
-                if let Some(o) = want {
-                    chosen.push(o);
-                }
-                for &c in choices {
-                    if chosen.len() >= *count {
-                        break;
-                    }
-                    if !chosen.contains(&c) {
-                        chosen.push(c);
-                    }
-                }
-                if probe
-                    .act(GameAction::SelectCards { cards: chosen })
-                    .is_err()
-                {
-                    break;
-                }
-            }
-            // CR 608.2d: accept a beneficial resolution-time "may" choice (the
-            // optional part of a "you may …" ability) so the loop's loop-closing
-            // action proceeds — e.g. Sword of the Paruns' "{3}: You may tap or
-            // untap equipped creature." Generalizes the optional-effect class.
-            WaitingFor::OptionalEffectChoice { .. } => {
-                if probe
-                    .act(GameAction::DecideOptionalEffect { accept: true })
-                    .is_err()
-                {
-                    break;
-                }
-            }
-            // CR 608.2d: a resolution-time "choose one of A or B" (e.g. a "tap or
-            // untap" ability, parsed to `Effect::ChooseOneOf`). Pick the branch
-            // whose effect is `SetTapState { Untap }`; fall back to the last branch
-            // if none match. Generalizes the modal-untap class, not one card.
-            WaitingFor::ChooseOneOfBranch { branches, .. } => {
-                use crate::types::ability::{Effect, TapStateChange};
-                let index = branches
-                    .iter()
-                    .position(|b| {
-                        matches!(
-                            *b.effect,
-                            Effect::SetTapState {
-                                state: TapStateChange::Untap,
-                                ..
-                            }
-                        )
-                    })
-                    .unwrap_or(branches.len().saturating_sub(1));
-                if probe.act(GameAction::ChooseBranch { index }).is_err() {
-                    break;
-                }
-            }
-            _ => break,
-        }
-    }
-}
-
-/// Run a combo driver: `setup` installs/attaches and returns the loop-step
-/// closure; the harness then warms up `WARMUP` cycles and measures up to `STEADY`
-/// steady cycles, returning the first confirmed certificate. The `step` closure
-/// drives exactly one loop iteration's actions.
-fn run_combo<S>(board: ComboBoard, mut step: S) -> Option<crate::analysis::LoopCertificate>
-where
-    S: FnMut(&mut LoopProbe),
-{
-    const WARMUP: usize = 2;
-    const STEADY: usize = 3;
-    let mut runner = board.runner;
-    let mut probe = LoopProbe::new(&mut runner);
-    for _ in 0..WARMUP {
-        step(&mut probe);
-        let _ = probe.iteration_delta();
-    }
-    for _ in 0..STEADY {
-        let start = probe.runner().state().clone();
-        // CR 606.3 / CR 704.5a: the loop's controller scopes the consumed-axis and
-        // win classification. Every combo scenario is built with the active player
-        // (P0) controlling the engine.
-        let controller = probe.runner().state().active_player;
-        step(&mut probe);
-        let delta = probe.iteration_delta();
-        let end = probe.runner().state().clone();
-        // Activated-ability loops are optional (CR 602.1), so `mandatory = false`.
-        if let Some(cert) = detect_loop(&start, &end, &delta, controller, false) {
-            return Some(cert);
-        }
-    }
-    None
-}
-
-// Effect predicates shared by drivers.
-fn is_mana_effect(e: &crate::types::ability::Effect) -> bool {
-    matches!(e, crate::types::ability::Effect::Mana { .. })
-}
-fn is_untap_effect(e: &crate::types::ability::Effect) -> bool {
-    use crate::types::ability::{Effect, TapStateChange};
-    matches!(
-        e,
-        Effect::SetTapState {
-            state: TapStateChange::Untap,
-            ..
-        }
-    )
-}
-/// An untap effect that untaps the SOURCE itself (`SelfRef`) — e.g. Staff of
-/// Domination's "{1}: Untap this artifact".
-fn is_self_untap_effect(e: &crate::types::ability::Effect) -> bool {
-    use crate::types::ability::{Effect, TapStateChange, TargetFilter};
-    matches!(
-        e,
-        Effect::SetTapState {
-            state: TapStateChange::Untap,
-            target: TargetFilter::SelfRef,
-            ..
-        }
-    )
-}
-/// An untap effect that untaps a *targeted* creature (a non-`SelfRef` filter) —
-/// e.g. Staff of Domination's "{3}, {T}: Untap target creature".
-fn is_target_creature_untap_effect(e: &crate::types::ability::Effect) -> bool {
-    use crate::types::ability::{Effect, TapStateChange, TargetFilter};
-    matches!(
-        e,
-        Effect::SetTapState {
-            state: TapStateChange::Untap,
-            target,
-            ..
-        } if !matches!(target, TargetFilter::SelfRef)
-    )
-}
-
-/// HELIOD, SUN-CROWNED + WALKING BALLISTA — the canonical driving combo, driven
-/// end-to-end through the real `apply()` pipeline with the cards' actual parsed
-/// abilities.
-///
-/// The repeating cycle (after the one-time lifelink grant, which is pre-loop
-/// setup, not part of the loop): Walking Ballista removes a +1/+1 counter to deal
-/// 1 damage to the opponent → lifelink gains its controller 1 life → Heliod's
-/// "whenever you gain life, put a +1/+1 counter on target creature you control"
-/// trigger returns the counter to Ballista. The board is identical at the end of
-/// each cycle; the monotone progress is +1 damage to the opponent and +1 life to
-/// the controller.
-///
-/// DISCRIMINATION: the `expect` flips if either `detect_loop` gate is reverted —
-/// the board returns identical only modulo the life/damage resources (board-
-/// equality gate) and the cycle's only net change is damage/life (net-progress
-/// gate). Skips (does not fail) if the export is unavailable.
-#[test]
-fn drive_heliod_ballista_certificate() {
-    let db = card_db();
-
-    let mut scenario = GameScenario::new();
-    scenario.at_phase(Phase::PreCombatMain);
-    scenario.with_life(P1, 40); // survive many pings within the test window
-    let mut runner = scenario.build();
-    {
-        let state = runner.state_mut();
-        state.active_player = P0;
-        state.priority_player = P0;
-    }
-
-    let (heliod, ballista) = {
-        let state = runner.state_mut();
-        let heliod = install_on_battlefield(state, db, "Heliod, Sun-Crowned", P0)
-            .expect("Heliod must be in the export");
-        let ballista = install_on_battlefield(state, db, "Walking Ballista", P0)
-            .expect("Walking Ballista must be in the export");
-        // Pre-loop setup (one-time, not part of the repeating cycle):
-        // (1) Ballista carries +1/+1 counters to remove (the loop refills them);
-        // (2) grant Ballista lifelink — in a real game Heliod's {1}{W} ability
-        //     does this once before the loop starts; we apply the resulting state
-        //     directly so the per-iteration cycle is exactly the repeating part.
-        {
-            let obj = state.objects.get_mut(&ballista).expect("ballista");
-            // Start with 2 counters: removing one to ping leaves Ballista a live
-            // 1/1 (not a 0/0 that dies to CR 704.5f before the Heliod trigger can
-            // return the counter), and Heliod's trigger refills it to 2 — so the
-            // board returns identical each cycle. This is the steady loop count
-            // the real combo settles into once life-gain replenishment matches the
-            // removal.
-            obj.counters
-                .insert(crate::types::counter::CounterType::Plus1Plus1, 2);
-            // Grant lifelink on the BASE keywords so `evaluate_layers` (which
-            // rebuilds `keywords` from `base_keywords` + layer effects) preserves
-            // it — pushing only onto `keywords` would be wiped by the recompute.
-            if !obj
-                .base_keywords
-                .contains(&crate::types::keywords::Keyword::Lifelink)
-            {
-                obj.base_keywords
-                    .push(crate::types::keywords::Keyword::Lifelink);
-                obj.keywords.push(crate::types::keywords::Keyword::Lifelink);
-            }
-        }
-        // CR 613: recompute layers so the granted keyword / counters take effect.
-        state.layers_dirty.mark_full();
-        crate::game::layers::evaluate_layers(state);
-        (heliod, ballista)
-    };
-    let _ = heliod;
-
-    // Find Ballista's "Remove a +1/+1 counter: deal 1 damage to any target"
-    // ability index (the one whose cost removes a counter), so the test does not
-    // hard-code an index that a card-data re-parse could reorder.
-    let remove_counter_idx = {
-        let obj = &runner.state().objects[&ballista];
-        obj.abilities
-            .iter()
-            .position(|a| matches!(*a.effect, Effect::DealDamage { .. }))
-            .expect("Ballista must have a deal-damage ability")
-    };
-
-    let mut probe = LoopProbe::new(&mut runner);
-
-    // WARMUP one full cycle to saturate per-turn bookkeeping (see the damage-loop
-    // test for the rationale), then compare two steady-state iterations.
-    drive_ballista_ping(&mut probe, ballista, remove_counter_idx);
-    let _ = probe.iteration_delta();
-
-    let cycle_start = probe.runner().state().clone();
-    drive_ballista_ping(&mut probe, ballista, remove_counter_idx);
-    let delta = probe.iteration_delta();
-    let cycle_end = probe.runner().state().clone();
-
-    let cert = detect_loop(&cycle_start, &cycle_end, &delta, P0, false).expect(
-        "Heliod + Ballista must be confirmed: board identical modulo life/damage, +1 damage/cycle",
-    );
-    assert_eq!(cert.win_kind, WinKind::LethalDamage);
-    assert!(
-        cert.covers(&[ResourceAxis::DamageDealt(P1)]),
-        "certificate must name unbounded damage to the opponent (got {:?})",
-        cert.unbounded
-    );
-}
-
-/// Drive one Walking Ballista "remove a +1/+1 counter: deal 1 to any target"
-/// activation at the opponent, then resolve everything on the stack (the ping AND
-/// the Heliod lifegain trigger that returns the counter).
-fn drive_ballista_ping(probe: &mut LoopProbe, ballista: ObjectId, ability_index: usize) {
-    let activated = probe
-        .act(GameAction::ActivateAbility {
-            source_id: ballista,
-            ability_index,
-        })
-        .expect("activate Ballista remove-counter ability");
-    // The ability targets "any target"; choose the opponent.
-    if matches!(activated.waiting_for, WaitingFor::TargetSelection { .. }) {
-        probe
-            .act(GameAction::SelectTargets {
-                targets: vec![TargetRef::Player(P1)],
-            })
-            .expect("target opponent with Ballista");
-    }
-    // CR 608: resolve the ping, then the Heliod lifegain trigger (which itself
-    // targets a creature you control — auto-resolved via choose_first_legal if it
-    // prompts). Pass priority / select trigger targets until the stack empties.
-    for _ in 0..16 {
-        if probe.runner().state().stack.is_empty()
-            && matches!(
-                probe.runner().state().waiting_for,
-                WaitingFor::Priority { .. }
-            )
-        {
-            break;
-        }
-        match &probe.runner().state().waiting_for {
-            WaitingFor::TargetSelection { .. } | WaitingFor::TriggerTargetSelection { .. } => {
-                // Heliod's counter-return trigger: target Ballista (a creature you
-                // control), so the board returns identical.
-                if probe
-                    .act(GameAction::SelectTargets {
-                        targets: vec![TargetRef::Object(ballista)],
-                    })
-                    .is_err()
-                {
-                    break;
-                }
-            }
-            _ => {
-                if probe.act(GameAction::PassPriority).is_err() {
-                    break;
-                }
-            }
-        }
-    }
-}
-
-// ===========================================================================
-// Per-combo bespoke drivers (real cards, real `apply()` pipeline). Each asserts
-// a confirmed `LoopCertificate` of the documented family + win_kind, and skips
-// (returns early) if the card-data export is absent.
-// ===========================================================================
-
-/// Assert a combo's driven certificate names the row's expected resource family
-/// and classifies the expected `win_kind`. For families whose concrete axis
-/// varies by card (mana color, which opponent, which counter class), this matches
-/// the family rather than one exact axis.
-fn assert_combo(idx: usize, cert: &crate::analysis::LoopCertificate) {
-    let row = &CORPUS[idx];
-    assert!(
-        cert.unbounded
-            .iter()
-            .any(|a| family_matches_axis(row.family, a)),
-        "{}: certificate {:?} must name a {:?}-family axis",
-        row.name,
-        cert.unbounded,
-        row.family,
-    );
-    assert_eq!(cert.win_kind, row.win_kind, "{}: win_kind", row.name);
-}
-
-/// Whether `axis` belongs to `family` (color/opponent/counter-class agnostic).
-fn family_matches_axis(family: ResourceFamily, axis: &ResourceAxis) -> bool {
-    use ResourceFamily as F;
-    match family {
-        F::Mana => matches!(axis, ResourceAxis::Mana(_)),
-        F::Tokens => matches!(axis, ResourceAxis::TokensCreated),
-        F::Damage | F::DrawDamage => matches!(axis, ResourceAxis::DamageDealt(_)),
-        F::Drain => matches!(axis, ResourceAxis::Life(_) | ResourceAxis::DamageDealt(_)),
-        F::Mill => matches!(axis, ResourceAxis::LibraryDelta(_)),
-        F::Death => matches!(
-            axis,
-            ResourceAxis::DeathTriggers
-                | ResourceAxis::SacTriggers
-                | ResourceAxis::LtbTriggers
-                | ResourceAxis::TokensCreated
-                | ResourceAxis::DamageDealt(_)
-                | ResourceAxis::Life(_)
-        ),
-        F::Landfall => matches!(
-            axis,
-            ResourceAxis::LandfallTriggers | ResourceAxis::EtbTriggers | ResourceAxis::Mana(_)
-        ),
-        F::Draw => matches!(
-            axis,
-            ResourceAxis::CardsDrawn | ResourceAxis::DamageDealt(_)
-        ),
-        F::Combat => matches!(axis, ResourceAxis::CombatPhases),
-        F::Turns => matches!(axis, ResourceAxis::ExtraTurns),
-        F::Counters => matches!(axis, ResourceAxis::Counter(_, _) | ResourceAxis::Life(_)),
-        F::Proliferate => matches!(axis, ResourceAxis::Trigger(_)),
-        F::Engine => true, // engine combos pump heterogeneous axes (mana/ETB/tokens/…)
-    }
-}
-
-/// #4 DEVOTED DRUID + VIZIER OF REMEDIES — infinite green mana.
-/// Cycle: tap Druid (add {G}); untap Druid (cost: put a -1/-1 counter on it, which
-/// Vizier reduces by one to *zero* counters — CR 614 replacement — so the untap is
-/// effectively free and no counter accrues). Board returns identical; +1 {G}/cycle.
-#[test]
-fn drive_combo_04_devoted_vizier() {
-    let board = build_board(CORPUS[6].cards);
-    let druid = board.ids[0];
-    let untap_idx = ability_index_where(board.runner.state(), druid, is_untap_effect)
-        .expect("Druid has an untap ability");
-    let cert = run_combo(board, |probe| {
-        // Untap Druid (free under Vizier), then tap it for {G}.
-        activate_and_resolve(probe, druid, untap_idx, None);
-        if let Some(tap_idx) = ability_index_where(probe.runner().state(), druid, is_mana_effect) {
-            activate_and_resolve(probe, druid, tap_idx, None);
-        }
-    })
-    .expect("Devoted Druid + Vizier must confirm infinite green mana");
-    assert_combo(6, &cert);
-}
-
-/// #2 GRIM MONOLITH + POWER ARTIFACT — infinite colorless mana.
-/// Power Artifact (Aura) enchants Grim Monolith, reducing its activated-ability
-/// costs by {2} (min 1), so the {4} untap becomes {2}. Cycle: tap Grim for {C}{C}
-/// {C} (+3); untap it for {2} (-2) → net +1 colorless/cycle, board identical.
-#[test]
-fn drive_combo_02_grim_power() {
-    let mut board = build_board(CORPUS[4].cards);
-    let grim = board.ids[0];
-    let power_artifact = board.ids[1];
-    attach_aura(board.runner.state_mut(), power_artifact, grim);
-    let untap_idx = ability_index_where(board.runner.state(), grim, is_untap_effect)
-        .expect("Grim Monolith has an untap ability");
-    let cert = run_combo(board, |probe| {
-        activate_and_resolve(probe, grim, untap_idx, None);
-        if let Some(tap_idx) = ability_index_where(probe.runner().state(), grim, is_mana_effect) {
-            activate_and_resolve(probe, grim, tap_idx, None);
-        }
-    })
-    .expect("Grim Monolith + Power Artifact must confirm infinite colorless mana");
-    assert_combo(4, &cert);
-}
-
-/// #47 SPIKE FEEDER + ARCHANGEL OF THUNE — infinite +1/+1 counters + life.
-/// Cycle: Spike Feeder removes a +1/+1 counter to gain 2 life; Archangel's
-/// "whenever you gain life, put a +1/+1 counter on each creature you control"
-/// returns a counter to Spike Feeder (net 0 on Spike) and pumps Archangel. The
-/// board is identical *modulo counters* (which the projection ignores); the
-/// unbounded axes are +1/+1 counters and life — `Counters` family.
-#[test]
-fn drive_combo_47_spike_archangel() {
-    let mut board = build_board(CORPUS[49].cards);
-    let spike = board.ids[0];
-    {
-        // CR 122: Spike Feeder "enters with two +1/+1 counters" — seed them (the
-        // as-enters replacement does not run for a directly-installed permanent).
-        let state = board.runner.state_mut();
-        if let Some(o) = state.objects.get_mut(&spike) {
-            o.counters
-                .insert(crate::types::counter::CounterType::Plus1Plus1, 2);
-        }
-        settle_layers(state);
-    }
-    let gain_idx = ability_index_where(board.runner.state(), spike, |e| {
-        matches!(e, crate::types::ability::Effect::GainLife { .. })
-    })
-    .expect("Spike Feeder has a remove-counter: gain-life ability");
-    let cert = run_combo(board, |probe| {
-        activate_and_resolve(probe, spike, gain_idx, None);
-    })
-    .expect("Spike Feeder + Archangel must confirm infinite counters + life");
-    assert_combo(49, &cert);
-}
-
-/// #7 BLOOM TENDER + FREED FROM THE REAL — infinite mana.
-/// Freed (Aura) enchants Bloom Tender, granting "{U}: untap enchanted creature".
-/// Bloom Tender taps for one mana of each color among permanents you control —
-/// green (Bloom Tender) + blue (Freed) = 2 mana. Cycle: tap Bloom Tender (+2),
-/// untap via Freed for {U} (-1) → net +1 mana/cycle, board identical.
-#[test]
-fn drive_combo_07_bloom_freed() {
-    let mut board = build_board(CORPUS[9].cards);
-    let bloom = board.ids[0];
-    let freed = board.ids[1];
-    attach_aura(board.runner.state_mut(), freed, bloom);
-    let untap_idx = ability_index_where(board.runner.state(), freed, is_untap_effect)
-        .expect("Freed from the Real has an untap ability");
-    let cert = run_combo(board, |probe| {
-        if let Some(tap_idx) = ability_index_where(probe.runner().state(), bloom, is_mana_effect) {
-            activate_and_resolve(probe, bloom, tap_idx, None);
-        }
-        activate_and_resolve(probe, freed, untap_idx, Some(TargetRef::Object(bloom)));
-    })
-    .expect("Bloom Tender + Freed must confirm infinite mana");
-    assert_combo(9, &cert);
-}
-
-/// #11 FAEBURROW ELDER + PEMMIN'S AURA — infinite mana (same family as Bloom +
-/// Freed): Pemmin's Aura grants "{U}: untap enchanted creature". Faeburrow taps
-/// for one mana of each color among your permanents (≥ 2). Cycle: tap Faeburrow
-/// (+N), untap via Pemmin for {U} (−1) → net (N − 1)/cycle, board identical.
-#[test]
-fn drive_combo_11_faeburrow_pemmin() {
-    let mut board = build_board(CORPUS[13].cards);
-    let faeburrow = board.ids[0];
-    let pemmin = board.ids[1];
-    attach_aura(board.runner.state_mut(), pemmin, faeburrow);
-    let pemmin_untap = ability_index_where(board.runner.state(), pemmin, is_untap_effect)
-        .expect("Pemmin's Aura has an untap ability");
-    let cert = run_combo(board, |probe| {
-        if let Some(tap_idx) =
-            ability_index_where(probe.runner().state(), faeburrow, is_mana_effect)
-        {
-            activate_and_resolve(probe, faeburrow, tap_idx, None);
-        }
-        activate_and_resolve(
-            probe,
-            pemmin,
-            pemmin_untap,
-            Some(TargetRef::Object(faeburrow)),
-        );
-    })
-    .expect("Faeburrow Elder + Pemmin's Aura must confirm infinite mana");
-    assert_combo(13, &cert);
-}
-
-/// #11 SELVALA, HEART OF THE WILDS + STAFF OF DOMINATION — infinite mana.
-/// Unlike the aura/self-untap mana combos above, the untap engine is a *second
-/// permanent's targeted untap*: Staff of Domination's "{3}, {T}: untap target
-/// creature" untaps Selvala, and "{1}: untap this artifact" untaps the Staff. With
-/// a high-power creature present, Selvala's "{G}, {T}: add X mana (X = greatest
-/// power among creatures you control)" reads X large, so one cycle is net mana-
-/// positive: tap Selvala for X, then untap Selvala ({3} + tap Staff) and untap
-/// Staff ({1}); X − {G} − {3} − {1} > 0 once X ≥ 6. Board returns identical (both
-/// untapped) modulo the floated mana. This exercises the multi-permanent untap
-/// chain the single-aura combos do not.
-#[test]
-fn drive_combo_11_selvala_staff() {
-    // 7/7 vanilla ⇒ greatest power = 7 ⇒ Selvala adds 7 mana, net +2/cycle.
-    let board = build_board_with_vanilla(CORPUS[12].cards, 7, 7);
-    let selvala = board.ids[0];
-    let staff = board.ids[1];
-    let selvala_tap = ability_index_where(board.runner.state(), selvala, is_mana_effect)
-        .expect("Selvala has a tap-for-mana ability");
-    let staff_untap_creature =
-        ability_index_where(board.runner.state(), staff, is_target_creature_untap_effect)
-            .expect("Staff of Domination has an untap-target-creature ability");
-    let staff_untap_self = ability_index_where(board.runner.state(), staff, is_self_untap_effect)
-        .expect("Staff of Domination has a self-untap ability");
-    let cert = run_combo(board, |probe| {
-        // Tap Selvala for X mana (X = greatest power), then untap her via Staff's
-        // targeted untap, then untap the Staff itself so it is ready next cycle.
-        activate_and_resolve(probe, selvala, selvala_tap, None);
-        activate_and_resolve(
-            probe,
-            staff,
-            staff_untap_creature,
-            Some(TargetRef::Object(selvala)),
-        );
-        activate_and_resolve(probe, staff, staff_untap_self, None);
-    })
-    .expect("Selvala + Staff of Domination must confirm infinite mana");
-    assert_combo(12, &cert);
-}
-
-/// D2 KILO, APOGEE MIND + FREED FROM THE REAL + RELIC OF LEGENDS — infinite
-/// proliferate triggers (mana-NEUTRAL — the canonical axis a mana-only model
-/// misses). Relic of Legends taps Kilo to add 1 mana; Kilo's "whenever Kilo
-/// becomes tapped, proliferate" fires; Freed (Aura on Kilo) untaps Kilo for {U}.
-/// Mana nets to zero (+1 Relic, -1 Freed); the only per-cycle progress is +1
-/// proliferate trigger. Board identical.
-#[test]
-fn drive_combo_d2_kilo_freed_relic() {
-    let mut board = build_board(CORPUS[1].cards);
-    let kilo = board.ids[0];
-    let freed = board.ids[1];
-    let relic = board.ids[2];
-    attach_aura(board.runner.state_mut(), freed, kilo);
-    // Relic's "tap a creature: add mana" ability (the one that taps Kilo) — found
-    // by its `TapCreatures` cost (Relic has two mana abilities; the tap-self one
-    // would not fire Kilo's trigger).
-    let relic_tap_creature = board.runner.state().objects[&relic]
-        .abilities
-        .iter()
-        .position(|a| {
-            matches!(
-                a.cost,
-                Some(crate::types::ability::AbilityCost::TapCreatures { .. })
-            )
-        })
-        .expect("Relic of Legends has a tap-a-creature mana ability");
-    let freed_untap = ability_index_where(board.runner.state(), freed, is_untap_effect)
-        .expect("Freed has an untap ability");
-    let cert = run_combo(board, |probe| {
-        // Tap Kilo via Relic's tap-a-creature cost (fires Kilo's "becomes tapped
-        // → proliferate" trigger), resolve, then untap Kilo via Freed.
-        activate_and_resolve(
-            probe,
-            relic,
-            relic_tap_creature,
-            Some(TargetRef::Object(kilo)),
-        );
-        activate_and_resolve(probe, freed, freed_untap, Some(TargetRef::Object(kilo)));
-    })
-    .expect("Kilo + Freed + Relic must confirm infinite proliferate triggers");
-    assert_combo(1, &cert);
-}
-
-/// Install `count` vanilla Elf creatures on P0's battlefield (CR 205.3: the
-/// "Elf" subtype is what an Elf-counting filter matches). Seeded directly onto
-/// the post-`build()` board (no enters trigger), mirroring `install_on_battlefield`
-/// but for an arbitrary-subtype vanilla. Settles layers so the count is live.
-fn seed_subtype_creatures(state: &mut GameState, subtype: &str, count: usize) {
-    use crate::types::card_type::CoreType;
-    use crate::types::identifiers::CardId;
-    use crate::types::zones::Zone;
-    for i in 0..count {
-        let card_id = CardId(state.next_object_id);
-        let id = crate::game::zones::create_object(
-            state,
-            card_id,
-            P0,
-            format!("{subtype} {i}"),
-            Zone::Battlefield,
-        );
-        if let Some(o) = state.objects.get_mut(&id) {
-            o.card_types.core_types.push(CoreType::Creature);
-            o.card_types.subtypes.push(subtype.to_string());
-            o.base_card_types = o.card_types.clone();
-            o.power = Some(1);
-            o.toughness = Some(1);
-            o.base_power = Some(1);
-            o.base_toughness = Some(1);
-            // CR 302.6: a pre-existing battlefield creature is not summoning-sick.
-            o.summoning_sick = false;
-        }
-    }
-    settle_layers(state);
-}
-
-/// #10 PRIEST OF TITANIA + UMBRAL MANTLE — infinite green mana.
-/// Priest taps for {G} per Elf on the battlefield; Umbral Mantle (Equipment)
-/// grants the equipped creature "{3}, {Q}: this creature gets +2/+2 until end of
-/// turn", whose {Q} cost untaps Priest. With ≥4 Elves (Priest itself is an Elf,
-/// plus seeded Elves), one cycle is net mana-positive: tap Priest for N green,
-/// untap via Umbral's {3}+{Q} ability (−{3} generic, paid from green) ⇒ net
-/// (N − 3) green/cycle. The +2/+2-until-end-of-turn buff climbs each cycle but is
-/// projected out (`project_out_resources` zeroes object power/toughness, and
-/// `GameState::PartialEq`/`objects_content_eq` ignore `transient_continuous_effects`
-/// / `next_continuous_effect_id`), so the board returns identical modulo the
-/// floated green. This exercises the equipment-granted modal-untap engine.
-#[test]
-fn drive_combo_10_priest_umbral() {
-    use crate::types::ability::Effect;
-    let mut board = build_board_green(CORPUS[10].cards);
-    let priest = board.ids[0];
-    let umbral = board.ids[1];
-    // 4 seeded Elves + Priest (itself an Elf) ⇒ Priest taps for 5 green; net +2 a
-    // cycle after the {3} untap cost.
-    seed_subtype_creatures(board.runner.state_mut(), "Elf", 4);
-    attach_aura(board.runner.state_mut(), umbral, priest);
-    let untap_idx = ability_index_where(board.runner.state(), priest, |e| {
-        matches!(e, Effect::Pump { .. })
-    })
-    .expect("Umbral Mantle grants Priest a {3},{Q} pump (untap) ability");
-    let cert = run_combo(board, |probe| {
-        if let Some(tap_idx) = ability_index_where(probe.runner().state(), priest, is_mana_effect) {
-            activate_and_resolve(probe, priest, tap_idx, None);
-        }
-        // Umbral's {3},{Q} ability: the {Q} cost untaps Priest for the next tap.
-        activate_and_resolve(probe, priest, untap_idx, None);
-    })
-    .expect("Priest of Titania + Umbral Mantle must confirm infinite green mana");
-    assert_combo(10, &cert);
-}
-
-/// REVERT-PROBE for [`drive_combo_10_priest_umbral`]: omit the Umbral untap step ⇒
-/// Priest stays tapped after the first cycle, so the second cycle's tap-for-mana
-/// fails and the board is NOT identical (Priest tapped, no further mana) ⇒
-/// `run_combo` finds no certificate. Proves the untap is load-bearing.
-#[test]
-fn drive_combo_10_priest_umbral_requires_untap() {
-    let mut board = build_board_green(CORPUS[10].cards);
-    let priest = board.ids[0];
-    let umbral = board.ids[1];
-    seed_subtype_creatures(board.runner.state_mut(), "Elf", 4);
-    attach_aura(board.runner.state_mut(), umbral, priest);
-    let cert = run_combo(board, |probe| {
-        if let Some(tap_idx) = ability_index_where(probe.runner().state(), priest, is_mana_effect) {
-            activate_and_resolve(probe, priest, tap_idx, None);
-        }
-        // No untap: Priest stays tapped, so this is not a repeatable loop.
-    });
-    assert!(
-        cert.is_none(),
-        "without the Umbral untap, Priest stays tapped — no loop"
-    );
-}
-
-/// #14 MARWYN, THE NURTURER + SWORD OF THE PARUNS — infinite green mana.
-/// Marwyn taps for {G} equal to her power; Sword of the Paruns (Equipment) grants
-/// "{3}: You may tap or untap equipped creature", whose untap mode readies Marwyn.
-/// With enough +1/+1 counters that her power exceeds {3}, one cycle is net mana-
-/// positive: tap Marwyn for P green, untap her via the Sword's {3} (paid from
-/// green) ⇒ net (P − 3) green/cycle, board identical modulo the floated green.
-/// The Sword's "{3}: You may tap or untap" parses to a `TargetOnly` ability whose
-/// `ChooseOneOf` sub-ability holds the real `SetTapState { Untap }` branch — the
-/// new `OptionalEffectChoice` (accept the "may") and `ChooseOneOfBranch` (pick the
-/// untap mode) auto-answers in `resolve_to_priority` drive it. This exercises the
-/// optional + modal untap engine the single-effect untap combos do not.
-#[test]
-fn drive_combo_14_marwyn_sword() {
-    use crate::types::ability::Effect;
-    let mut board = build_board_green(CORPUS[14].cards);
-    let marwyn = board.ids[0];
-    let sword = board.ids[1];
-    // CR 122: Marwyn's base power is 1; seed 6 +1/+1 counters (power 7) so a cycle
-    // nets +4 green after the {3} untap cost. The "another Elf enters" trigger does
-    // not fire for directly-installed permanents, so seed the counters directly.
-    {
-        let state = board.runner.state_mut();
-        if let Some(o) = state.objects.get_mut(&marwyn) {
-            o.counters
-                .insert(crate::types::counter::CounterType::Plus1Plus1, 6);
-        }
-        settle_layers(state);
-    }
-    attach_aura(board.runner.state_mut(), sword, marwyn);
-    let sword_modal = ability_index_where(board.runner.state(), sword, |e| {
-        matches!(e, Effect::TargetOnly { .. })
-    })
-    .expect("Sword of the Paruns has a {3}: tap-or-untap modal ability");
-    let cert = run_combo(board, |probe| {
-        if let Some(tap_idx) = ability_index_where(probe.runner().state(), marwyn, is_mana_effect) {
-            activate_and_resolve(probe, marwyn, tap_idx, None);
-        }
-        // Sword's {3} modal: accept the "you may" + choose the untap branch
-        // (auto-answered), targeting Marwyn (the equipped creature) so she readies.
-        activate_and_resolve(probe, sword, sword_modal, Some(TargetRef::Object(marwyn)));
-    })
-    .expect("Marwyn + Sword of the Paruns must confirm infinite green mana");
-    assert_combo(14, &cert);
-}
-
-/// REVERT-PROBE for [`drive_combo_14_marwyn_sword`]: omit the Sword untap step ⇒
-/// Marwyn stays tapped, the next cycle's tap-for-mana fails, and the board is not
-/// identical ⇒ no certificate. Proves the modal untap is load-bearing.
-#[test]
-fn drive_combo_14_marwyn_sword_requires_untap() {
-    let mut board = build_board_green(CORPUS[14].cards);
-    let marwyn = board.ids[0];
-    let sword = board.ids[1];
-    {
-        let state = board.runner.state_mut();
-        if let Some(o) = state.objects.get_mut(&marwyn) {
-            o.counters
-                .insert(crate::types::counter::CounterType::Plus1Plus1, 6);
-        }
-        settle_layers(state);
-    }
-    attach_aura(board.runner.state_mut(), sword, marwyn);
-    let cert = run_combo(board, |probe| {
-        if let Some(tap_idx) = ability_index_where(probe.runner().state(), marwyn, is_mana_effect) {
-            activate_and_resolve(probe, marwyn, tap_idx, None);
-        }
-        // No untap: Marwyn stays tapped, so this is not a repeatable loop.
-    });
-    assert!(
-        cert.is_none(),
-        "without the Sword untap, Marwyn stays tapped — no loop"
-    );
-}
-
 /// END-TO-END DRIVING TEST (the Heliod-shaped damage acceptance): a board-neutral
 /// repeatable ping loop, driven through the real `apply()` pipeline via
 /// [`LoopProbe`], is confirmed by [`detect_loop`] as a `LethalDamage` net-progress
 /// loop whose unbounded axis is damage to the opponent.
 ///
 /// DISCRIMINATION: this is the assertion that flips if the detector is reverted.
-/// - Revert the `is_net_progress` gate ⇒ `detect_loop` returns `None` here even
-///   though damage advanced (so the `expect` panics).
-/// - Revert the `loop_states_equal_modulo_resources` gate ⇒ the two boundary
-///   states (which differ only by the opponent's life) compare unequal and
-///   `detect_loop` returns `None` (so the `expect` panics).
-///
-/// The negative control is `drive_board_change_is_not_a_loop` below.
 #[test]
 fn drive_damage_loop_certificate() {
     let (scenario, pinger) = pinger_scenario(1);
@@ -1775,12 +621,8 @@ fn drive_damage_loop_certificate() {
 
     let mut probe = crate::analysis::LoopProbe::new(&mut runner);
 
-    // WARMUP: drive one full ping cycle first. This saturates the per-turn
-    // bookkeeping a first activation perturbs (e.g. `objects_that_dealt_damage`
-    // gains the pinger, the activated-ability tally increments) so that two
-    // *subsequent* steady-state iterations differ only by the monotone resource
-    // (the opponent's life / the damage tally), which is exactly the loop point a
-    // CR 732.2a shortcut compares.
+    // WARMUP: drive one full ping cycle first to saturate per-turn bookkeeping so
+    // two subsequent steady-state iterations differ only by the monotone resource.
     drive_one_ping(&mut probe, pinger);
     let _ = probe.iteration_delta(); // discard warmup; roll the boundary forward
 
@@ -1801,11 +643,8 @@ fn drive_damage_loop_certificate() {
     );
 }
 
-/// END-TO-END SOUNDNESS NEGATIVE: a single ping is genuinely board-identical, but
-/// if we instead drive an action that changes the board (cast a creature from
-/// hand to the battlefield), the start/end states differ structurally and
-/// [`detect_loop`] must return `None` — no false certificate. This is the
-/// revert-probe for the board-equality gate at the *driven* level.
+/// END-TO-END SOUNDNESS NEGATIVE: an action that changes the board (cast a
+/// creature from hand) yields differing start/end states ⇒ no certificate.
 #[test]
 fn drive_board_change_is_not_a_loop() {
     let mut scenario = GameScenario::new();
@@ -1856,9 +695,7 @@ fn drive_board_change_is_not_a_loop() {
 }
 
 /// END-TO-END SOUNDNESS NEGATIVE: an idle board with NO driven progress yields no
-/// certificate. Snapshot the same state twice with an empty event tally: the
-/// board is identical but nothing advanced, so `detect_loop` returns `None`.
-/// Revert-probe for the net-progress gate at the driven level.
+/// certificate. Revert-probe for the net-progress gate at the driven level.
 #[test]
 fn drive_idle_board_is_not_a_loop() {
     let (scenario, _pinger) = pinger_scenario(1);
@@ -1875,233 +712,34 @@ fn drive_idle_board_is_not_a_loop() {
     );
 }
 
-// ---------------------------------------------------------------------------
-// Remaining corpus rows: documented (non-driven) data rows.
-//
-// The combos NOT driven above fall into structural buckets that PR-2's detection
-// model or a per-combo driver does not yet reach. Each bucket below is a PRECISE,
-// measured reason an honest driver cannot confirm the loop on today's engine —
-// not a TODO. The drivable shape PR-2 reaches is exactly the IN-PLACE loop: the
-// same objects tap/untap or gain/lose counters/P-T each cycle, with no object
-// leaving and re-entering. A continuous (non-counter) P/T buff is NOT a blocker:
-// `project_out_resources` sets each object's `power`/`toughness` to `None`, and
-// `GameState::PartialEq` ignores `transient_continuous_effects` /
-// `next_continuous_effect_id`, so the climbing buff is projected out (this is why
-// Priest of Titania + Umbral Mantle is driven — its untap rides a
-// "{3},{Q}: +2/+2 until end of turn" granted ability).
-//
-//  * OBJECT RE-ENTRY loops (tokens / blink / persist / undying / recur engines:
-//    Kiki-Jiki, Splinter Twin, Midnight Guard, Scurry Oak, Felidar Guardian,
-//    Mikaeus + Triskelion, Gravecrawler, Karmic Guide, Reassembling Skeleton,
-//    Earthcraft, Palinchron + Deadeye, Dockside + Temur, Food Chain, …). A
-//    permanent that dies/blinks/bounces and returns gets a FRESH `ObjectId`
-//    every cycle, so the id-keyed per-object lookup in
-//    `game_state::objects_content_eq` misses (returns `None`) and the
-//    `battlefield` order changes — `loop_states_equal_modulo_resources`
-//    (via `objects_content_eq` / `PartialEq`) sees a different board.
-//    Confirming these needs an object-identity-canonicalizing projection, a
-//    follow-up beyond PR-2's "board identical modulo monotone resources" model.
-//  * EXTRA-TURN / EXTRA-COMBAT loops (Time Sieve, Aggravated Assault, Combat
-//    Celebrant): each cycle advances `turn_number` / combat count and re-enters
-//    phases, so the loop point is a different turn/phase — not board-identical.
-//  * COLOR-CONVERTING net-progress that the per-color rule rejects (Pili-Pala +
-//    Grand Architect): the engine models mana per color; a producer that gains
-//    one color while a *different* color (or restricted mana) it cannot replace
-//    is consumed reads as a per-color deficit, which `loop_check::is_progress`
-//    rightly rejects. (Selvala + Staff IS driven because Selvala can produce the
-//    one color its whole cycle consumes — see that driver's green-only float.)
-//  * DRAIN FEEDBACK cascades (Sanguine Bond + Exquisite Blood [idx 17]; Marauding
-//    Blight-Priest + Bloodthirsty Conqueror [idx 18]): a mandatory triggered cascade
-//    seeded by a clean external life-gain. PR-3 (Option C) DRIVES BOTH LIVE — they are
-//    promoted to `DRIVEN_ROW_INDICES`. The persisted `loop_detect_ring` accumulates
-//    across the per-beat `apply(PassPriority)` drive and `reconcile_terminal_result`'s
-//    §3 shortcut emits `GameOver{Some(P0)}` by ~beat 6 (CR 732.2a → CR 704.5a), well
-//    before the high-life 704.5a death. idx 17's targeted "target opponent loses life"
-//    trigger auto-resolves to the sole legal target (the opponent — "target opponent"
-//    has exactly one legal target in two-player; no target-selection stop), so the targeted
-//    shape drives identically. (Drivers: `drive_drain_idx18_wins_live`,
-//    `drive_drain_idx17_targeted_wins_live`; recursion-regression backstop
-//    `drive_drain_idx18_legal_actions_terminates_bounded`.)
-//  * CARD-GATED (4): Doc Aurlock / Professor Onyx / Animate Dead / Grindstone +
-//    Painter's Servant have Unimplemented parts (§3).
-//
-// Every remaining row stays in `CORPUS` (shape-locked by
-// `corpus_table_shape_is_locked`) and its cards are validated present + (for the
-// 49 non-gated) implemented by `corpus_cards_present_and_implementation_status_
-// matches_gating`. A follow-up (or PR-5's `combo-verify` CLI) adds the bespoke
-// drivers; the detector each would exercise is already covered by the driven
-// combos above and the `loop_check.rs` building-block tests.
-// ---------------------------------------------------------------------------
-
-/// The corpus rows confirmed end-to-end by a `drive_combo_*` / `drive_*` test in
-/// this module (real cards through the real `apply()` pipeline). Locked here so a
-/// regression that silently stops driving a combo is caught by
-/// `confirmed_drivers_match_expected`.
-const DRIVEN_ROW_INDICES: &[usize] = &[
-    0,  // Heliod, Sun-Crowned + Walking Ballista  (drive_heliod_ballista_certificate)
-    1,  // Kilo + Freed + Relic                     (drive_combo_d2_kilo_freed_relic)
-    4,  // Grim Monolith + Power Artifact           (drive_combo_02_grim_power)
-    6,  // Devoted Druid + Vizier of Remedies       (drive_combo_04_devoted_vizier)
-    9,  // Bloom Tender + Freed from the Real       (drive_combo_07_bloom_freed)
-    10, // Priest of Titania + Umbral Mantle        (drive_combo_10_priest_umbral)
-    12, // Selvala, Heart of the Wilds + Staff      (drive_combo_11_selvala_staff)
-    13, // Faeburrow Elder + Pemmin's Aura          (drive_combo_11_faeburrow_pemmin)
-    14, // Marwyn, the Nurturer + Sword of Paruns   (drive_combo_14_marwyn_sword)
-    17, // Sanguine Bond + Exquisite Blood          (drive_drain_idx17_targeted_wins_live)
-    18, // Marauding Blight-Priest + Conqueror      (drive_drain_idx18_wins_live)
-    49, // Spike Feeder + Archangel of Thune        (drive_combo_47_spike_archangel)
-];
-
-/// Meta: the driven set is a subset of the non-gated corpus and every driven row
-/// is currently driven by a real test (kept honest as drivers are added/removed).
-#[test]
-fn confirmed_drivers_match_expected() {
-    for &idx in DRIVEN_ROW_INDICES {
-        assert!(idx < CORPUS.len(), "driven index in range");
-        assert!(
-            CORPUS[idx].gated_on.is_none(),
-            "{}: a driven combo must not be card-gated",
-            CORPUS[idx].name
-        );
-    }
-}
-
 // ===========================================================================
-// PR-3 (Option C) live drain-cascade drivers. These drive the REAL per-beat
+// PR-3 (Option C) live drain-cascade tests. These drive the REAL per-beat
 // `apply(PassPriority)` reducer (not the offline `detect_loop` harness) so the
-// persisted `loop_detect_ring` accumulation (§2) and the reconcile-seam win
-// shortcut (§3) are exercised end-to-end. Each names its revert-fail line.
+// persisted `loop_detect_ring` accumulation and the reconcile-seam win shortcut
+// are exercised end-to-end. Each names its revert-fail line. Board install +
+// cascade seeding + per-beat drive use the shared `corpus::` toolkit.
 // ===========================================================================
-
-/// Build a 2-player board with the named permanents installed on P0, P1 set to a
-/// high life total `victim_life` (so a natural CR 704.5a death cannot be the cause
-/// of any early `GameOver`), and the active player P0 at a clean `PreCombatMain`
-/// priority window. No mana is floated (the drain cascade is trigger-driven, not
-/// activated). Returns `None` if the export is absent or any name is missing.
-fn build_drain_board(cards: &[&str], victim_life: i32) -> Option<ComboBoard> {
-    let db = card_db();
-    if cards.iter().any(|c| db.get_face_by_name(c).is_none()) {
-        return None;
-    }
-    let mut scenario = GameScenario::new();
-    scenario.at_phase(Phase::PreCombatMain);
-    scenario.with_life(P0, 40);
-    scenario.with_life(P1, victim_life);
-    let mut runner = scenario.build();
-    {
-        let state = runner.state_mut();
-        state.active_player = P0;
-        state.priority_player = P0;
-    }
-    let mut ids = Vec::new();
-    {
-        let state = runner.state_mut();
-        for &name in cards {
-            ids.push(install_on_battlefield(state, db, name, P0)?);
-        }
-        settle_layers(state);
-    }
-    Some(ComboBoard { runner, ids })
-}
-
-/// Seed a drain cascade by gaining P0 1 life through the real life-gain pipeline
-/// and placing the resulting "whenever you gain life" trigger on the stack via the
-/// production trigger chokepoint (`process_triggers`). Leaves the board at a
-/// priority window with exactly one trigger on the stack — the same shape a real
-/// external lifegain would produce. Returns the stack length after seeding.
-fn seed_lifegain_cascade(board: &mut ComboBoard) -> usize {
-    let state = board.runner.state_mut();
-    let mut events = Vec::new();
-    // CR 119.3: gain P0 1 life — fires Marauding Blight-Priest's "whenever you gain
-    // life" trigger.
-    let _ = crate::game::effects::life::apply_life_gain(state, P0, 1, &mut events);
-    // CR 603.3: put the triggered ability on the stack as a player would receive
-    // priority — the production placement path.
-    crate::game::triggers::process_triggers(state, &events);
-    // Reset to a clean active-player priority window (the seed is pre-loop setup).
-    state.priority_player = state.active_player;
-    state.waiting_for = WaitingFor::Priority {
-        player: state.active_player,
-    };
-    state.stack.len()
-}
-
-/// One observation of the live per-beat drive.
-#[derive(Debug)]
-struct BeatTrace {
-    beat: usize,
-    wf: WaitingFor,
-    stack_len: usize,
-    ring_len: usize,
-    p0_life: i32,
-    p1_life: i32,
-}
-
-/// Drive `runner.act(PassPriority)` up to `max_beats` times, recording one
-/// [`BeatTrace`] per beat. Stops early on `GameOver`. Returns the trace.
-fn drive_pass_priority(board: &mut ComboBoard, max_beats: usize) -> Vec<BeatTrace> {
-    let mut trace = Vec::new();
-    for beat in 1..=max_beats {
-        if matches!(
-            board.runner.state().waiting_for,
-            WaitingFor::GameOver { .. }
-        ) {
-            break;
-        }
-        if board.runner.act(GameAction::PassPriority).is_err() {
-            break;
-        }
-        let s = board.runner.state();
-        trace.push(BeatTrace {
-            beat,
-            wf: s.waiting_for.clone(),
-            stack_len: s.stack.len(),
-            ring_len: s.loop_detect_ring.len(),
-            p0_life: s.players[0].life,
-            p1_life: s.players[1].life,
-        });
-        if matches!(
-            board.runner.state().waiting_for,
-            WaitingFor::GameOver { .. }
-        ) {
-            break;
-        }
-    }
-    trace
-}
-
-/// Index of the first beat whose `waiting_for` is `GameOver { winner: Some(w) }`,
-/// or `None` if no early win occurred within the driven window.
-fn first_gameover_beat(trace: &[BeatTrace]) -> Option<(usize, PlayerId)> {
-    trace.iter().find_map(|t| match t.wf {
-        WaitingFor::GameOver {
-            winner: Some(winner),
-        } => Some((t.beat, winner)),
-        _ => None,
-    })
-}
 
 /// C-L1: the PERSISTED loop-detection ring wins idx 18 (Marauding Blight-Priest +
 /// Bloodthirsty Conqueror) LIVE under the default per-beat `apply(PassPriority)`
-/// drive. P1 starts at 200 life so a natural CR 704.5a death cannot be the cause
-/// (it would take ~400 beats); the live `GameOver{Some(P0)}` fires by ~beat 6 from
-/// the accumulated ring + the §3 reconcile shortcut.
+/// drive. P1 starts at 200 life so a natural CR 704.5a death cannot be the cause;
+/// the live `GameOver{Some(P0)}` fires by ~beat 6 from the accumulated ring + the
+/// §3 reconcile shortcut.
 ///
 /// REVERT-FAIL (the named discriminators, each flips this assertion):
-///  (a) remove the §3 block in `reconcile_terminal_result` (engine.rs) ⇒ the
-///      cascade grinds, no early `GameOver` within the window ⇒ `expect` fails.
+///  (a) remove the §3 block in `reconcile_terminal_result` ⇒ the cascade grinds,
+///      no early `GameOver` ⇒ `expect` fails.
 ///  (b) remove the relocated §2 `record_loop_detect_sample` block in
-///      `pass_priority_once_with_pipeline` ⇒ the ring never persists across beats
-///      (caps at 0), §3 never matches ⇒ no early `GameOver` ⇒ `expect` fails. This
-///      proves the PERSISTED ring is the load-bearing new surface.
+///      `pass_priority_once_with_pipeline` ⇒ the ring never persists across beats ⇒
+///      no early `GameOver` ⇒ `expect` fails.
 #[test]
 fn drive_drain_idx18_wins_live() {
-    let Some(mut board) = build_drain_board(CORPUS[18].cards, 200) else {
+    let Some(mut board) = corpus::build_drain_board(card_db(), corpus::row(18).cards, 200) else {
         return; // export absent (CI / fresh checkout): skip, never fail spuriously
     };
-    seed_lifegain_cascade(&mut board);
-    let trace = drive_pass_priority(&mut board, 40);
-    let (beat, winner) = first_gameover_beat(&trace)
+    corpus::seed_lifegain_cascade(&mut board);
+    let trace = corpus::drive_pass_priority(&mut board, 40);
+    let (beat, winner) = corpus::first_gameover_beat(&trace)
         .expect("idx18 drain cascade must win LIVE via the persisted ring + §3 shortcut");
     assert_eq!(
         winner, P0,
@@ -2112,8 +750,7 @@ fn drive_drain_idx18_wins_live() {
         "the live win must fire from the ring (~beat 6), not the ~400-beat 704.5a death; got beat {beat}"
     );
     // The PERSISTED ring accumulated across beats (≥ 3 snapshots are needed for the
-    // modulo-match that fires the win — see the §4 trace), and the cascade held the
-    // stack non-empty the whole time (a self-refilling, NON-shrinking loop).
+    // modulo-match that fires the win), and the cascade held the stack non-empty.
     let max_ring = trace.iter().map(|t| t.ring_len).max().unwrap_or(0);
     assert!(
         max_ring >= 3,
@@ -2138,27 +775,19 @@ fn drive_drain_idx18_wins_live() {
 }
 
 /// C-L2 (a genuinely SECOND loop shape — idx 17, Sanguine Bond + Exquisite Blood,
-/// the TARGETED `LoseLife` variant). §4 disposition measurement: the per-beat drive
-/// auto-resolves Sanguine Bond's "target opponent loses that much life" trigger to
-/// the sole legal target (opponent P1) with NO target-selection stop — "target
-/// opponent" has exactly one legal target in a two-player game, so this targeted
-/// shape also wins live — promoting idx 17 alongside idx 18. The assertion that NO
-/// `TargetSelection`/`TriggerTargetSelection` window ever appears is what makes the
-/// auto-resolution claim non-vacuous (if the engine stopped for a target, this test
-/// would catch it and idx 17 would stay targeting-deferred).
+/// the TARGETED `LoseLife` variant). The per-beat drive auto-resolves the "target
+/// opponent loses that much life" trigger to the sole legal target (opponent P1)
+/// with NO target-selection stop, so this targeted shape also wins live.
 ///
 /// REVERT-FAIL: same as C-L1 — removing the §3 block or the §2 sample drops the
-/// early `GameOver`. Additionally exercises §7's stack-id canon
-/// (`project_out_resources`, resource.rs:751): reverting the `entry.id =
-/// ObjectId(pos)` canon loop makes each refilled trigger's fresh `ObjectId` defeat
-/// the modulo-match, so no cycle point is ever found ⇒ no early `GameOver`.
+/// early `GameOver`.
 #[test]
 fn drive_drain_idx17_targeted_wins_live() {
-    let Some(mut board) = build_drain_board(CORPUS[17].cards, 200) else {
+    let Some(mut board) = corpus::build_drain_board(card_db(), corpus::row(17).cards, 200) else {
         return; // export absent: skip
     };
-    seed_lifegain_cascade(&mut board);
-    let trace = drive_pass_priority(&mut board, 40);
+    corpus::seed_lifegain_cascade(&mut board);
+    let trace = corpus::drive_pass_priority(&mut board, 40);
     // The targeted trigger auto-resolves: no target window appears at any beat.
     assert!(
         trace.iter().all(|t| !matches!(
@@ -2167,7 +796,7 @@ fn drive_drain_idx17_targeted_wins_live() {
         )),
         "idx17's targeted trigger must auto-resolve (no target-selection stop) under the per-beat drive"
     );
-    let (beat, winner) = first_gameover_beat(&trace)
+    let (beat, winner) = corpus::first_gameover_beat(&trace)
         .expect("idx17 targeted drain cascade must win LIVE (auto-resolved target)");
     assert_eq!(winner, P0);
     assert!(
@@ -2182,40 +811,20 @@ fn drive_drain_idx17_targeted_wins_live() {
 }
 
 /// C-L1-probe — Defect-2 BOUNDED-TERMINATION regression guard. After the live drive
-/// has populated the ring (stack≠∅, at the RESOLVING `Priority{non-active}` window,
-/// ring ≥ 2, BEFORE any `GameOver`), call `legal_actions` directly. The legality
+/// has populated the ring at the RESOLVING `Priority{non-active}` window (ring ≥ 2,
+/// stack≠∅, BEFORE any `GameOver`), call `legal_actions` directly. The legality
 /// probe clones-and-applies `PassPriority`, which resolves the top and re-enters
 /// `reconcile_terminal_result` §3 — the seam the Defect-2 recursion concern is about.
-/// This asserts the call TERMINATES with a bounded, non-empty action list and does
-/// not mutate the live game.
-///
-/// HONEST DISCRIMINATION NOTE (measured, not assumed): in the SHIPPED architecture
-/// this test is a bounded-termination *property* test, NOT a non-vacuous discriminator
-/// of the `&& !in_simulation_probe()` guard. Measured across three revert configs
-/// (drop the §3 guard; drop the filter.rs `SimulationProbeGuard::enter()` so the flag
-/// is never set; both with the real per-beat drive AND this direct `legal_actions`
-/// call), the `reconcile→§3→§9→legal_actions→SimulationFilter→reconcile` path does NOT
-/// recurse — the §9 gate (`no_living_player_has_meaningful_priority_action`) resets
-/// each probe's priority/pass state, so the nested `SimulationFilter` applies are
-/// handoffs that never re-resolve, and §3 (which only fires when a winner is FOUND)
-/// completes at ring=3 on the real path before the ring could grow. The thread-local
-/// guard is therefore DEFENSIVE DEPTH enforcing the top-level-only invariant, not the
-/// "sole barrier" the plan framed; the original SIGABRT (impl-report §2, observed at
-/// ring cap 16 with a different §9/§2 configuration) does not reproduce here. This
-/// test still guards against a FUTURE change that drops the §9 reset and re-opens the
-/// recursion: with such a change AND the guard removed it would overflow; with the
-/// guard it stays bounded.
+/// This asserts the call TERMINATES with a bounded, non-empty action list.
 #[test]
 fn drive_drain_idx18_legal_actions_terminates_bounded() {
-    let Some(mut board) = build_drain_board(CORPUS[18].cards, 200) else {
+    let Some(mut board) = corpus::build_drain_board(card_db(), corpus::row(18).cards, 200) else {
         return; // export absent: skip
     };
-    seed_lifegain_cascade(&mut board);
+    corpus::seed_lifegain_cascade(&mut board);
     // Drive until the ring is populated (≥ 2 snapshots) at the RESOLVING priority
     // window — `Priority{non-active}` (the last passer), where the NEXT all-pass
-    // resolves the top and would push the ring to a modulo MATCH. This is the exact
-    // state whose legality probe (a clone-and-apply of `PassPriority`) re-enters
-    // `reconcile_terminal_result` §3; stop here, BEFORE the real drive fires GameOver.
+    // resolves the top and would push the ring to a modulo MATCH.
     let active = board.runner.state().active_player;
     let mut reached = false;
     for _ in 0..40 {
@@ -2259,8 +868,7 @@ fn drive_drain_idx18_legal_actions_terminates_bounded() {
 }
 
 /// Install a board-neutral artifact on `player` with a costless "draw 1" activated
-/// ability — a meaningful (loop-ending) priority action. Mirrors the engine's
-/// `add_non_mana_activated_artifact` U-gate helper at the corpus level.
+/// ability — a meaningful (loop-ending) priority action.
 fn add_meaningful_action_artifact(state: &mut GameState, player: PlayerId) -> ObjectId {
     use crate::types::card_type::CoreType;
     use crate::types::identifiers::CardId;
@@ -2286,47 +894,39 @@ fn add_meaningful_action_artifact(state: &mut GameState, player: PlayerId) -> Ob
 }
 
 /// C-neg-A (+ the live face of the all-players §9 probe): a victim that holds a
-/// MEANINGFUL loop-ending action is NEVER shortcut-eliminated. The victim P1 (the
-/// NON-current holder at the §3 site, where priority has reset to active P0) holds
+/// MEANINGFUL loop-ending action is NEVER shortcut-eliminated. The victim P1 holds
 /// a costless "draw 1" ability, so `no_living_player_has_meaningful_priority_action`
-/// probes P1, finds the out, and returns `false` ⇒ the §3 shortcut is refused and
-/// the cascade falls through to the existing grind — no early `GameOver`.
+/// finds the out ⇒ the §3 shortcut is refused and no early `GameOver` fires.
 ///
 /// REVERT-FAIL:
 ///  - remove the §9 gate call at the §3 site ⇒ `GameOver{Some(P0)}` fires while P1
 ///    had an out (unsound) — this assertion (`no early GameOver`) flips.
-///  - swap §9 for the current-holder-only `priority_player_has_meaningful_action`
-///    (which probes only P0, who has no action) ⇒ it misses P1's masked out and the
-///    shortcut wrongly fires (the all-players generalization is load-bearing; the
-///    unit-level `loop_gate_probes_all_living_players_not_just_current_holder`
-///    pins the same distinction).
+///  - swap §9 for the current-holder-only probe ⇒ it misses P1's masked out and the
+///    shortcut wrongly fires.
 #[test]
 fn drive_drain_idx18_victim_with_out_is_not_eliminated() {
-    let Some(mut board) = build_drain_board(CORPUS[18].cards, 200) else {
+    let Some(mut board) = corpus::build_drain_board(card_db(), corpus::row(18).cards, 200) else {
         return; // export absent: skip
     };
     add_meaningful_action_artifact(board.runner.state_mut(), P1);
-    settle_layers(board.runner.state_mut());
-    seed_lifegain_cascade(&mut board);
-    let trace = drive_pass_priority(&mut board, 40);
+    corpus::settle_layers(board.runner.state_mut());
+    corpus::seed_lifegain_cascade(&mut board);
+    let trace = corpus::drive_pass_priority(&mut board, 40);
     assert!(
-        first_gameover_beat(&trace).is_none(),
+        corpus::first_gameover_beat(&trace).is_none(),
         "P1 holds a loop-ending action — the §9 gate must refuse the shortcut (no GameOver)"
     );
 }
 
 /// C-neg-D — ring HYGIENE: a finite, non-refilling multi-spell stack that drains to
 /// empty NEVER accumulates loop snapshots and NEVER produces a `GameOver`. Three
-/// costless "draw 1" abilities are stacked, then resolved one-per-beat; each
-/// resolution SHRINKS the stack (`len_after < len_before`), so the §2 refill gate's
-/// `stack.len() >= stack_len_before` clause is false ⇒ the clear arm runs ⇒ the
-/// ring stays empty after every shrinking resolution.
+/// costless "gain 1 life" abilities are stacked, then resolved one-per-beat; each
+/// resolution SHRINKS the stack, so the §2 refill gate's `stack.len() >=
+/// stack_len_before` clause is false ⇒ the clear arm runs ⇒ the ring stays empty.
 ///
 /// REVERT-FAIL: removing the `state.stack.len() >= stack_len_before` clause from the
 /// §2 refill gate makes a normal shrinking resolution RECORD a snapshot ⇒ the ring
-/// becomes non-empty ⇒ the `is_empty()` assertion flips. (No `GameOver` either way —
-/// these resolutions drain no life — so the ring-emptiness assertion is the
-/// load-bearing hygiene discriminator.)
+/// becomes non-empty ⇒ the `is_empty()` assertion flips.
 #[test]
 fn drive_finite_stack_keeps_ring_empty() {
     let mut scenario = GameScenario::new();
@@ -2366,8 +966,8 @@ fn drive_finite_stack_keeps_ring_empty() {
         ));
         id
     };
-    settle_layers(runner.state_mut());
-    let gain_idx = ability_index_where(runner.state(), artifact, |e| {
+    corpus::settle_layers(runner.state_mut());
+    let gain_idx = corpus::ability_index_where(runner.state(), artifact, |e| {
         matches!(e, Effect::GainLife { .. })
     })
     .expect("the artifact has a gain-life ability");
@@ -2382,10 +982,9 @@ fn drive_finite_stack_keeps_ring_empty() {
             .expect("activate costless gain-life");
     }
     assert_eq!(runner.state().stack.len(), 3, "three abilities stacked");
-    // Drive PassPriority ONLY until the finite stack drains to empty (driving past
-    // empty would advance turns into a draw-step deck-out — unrelated to loop
-    // hygiene). The ring must stay empty at EVERY shrinking-resolution beat, and no
-    // shortcut GameOver may occur while the stack is non-empty.
+    // Drive PassPriority ONLY until the finite stack drains to empty. The ring must
+    // stay empty at EVERY shrinking-resolution beat, and no shortcut GameOver may
+    // occur while the stack is non-empty.
     let mut resolutions = 0;
     let mut max_ring = 0usize;
     for _ in 0..20 {
